% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_parameters.R
\name{adjust_param}
\alias{adjust_param}
\alias{adjust_shape}
\alias{adjust_scale}
\alias{adjust_kappa}
\title{Adjust parameters}
\usage{
adjust_shape(tentative, modifier = 0)

adjust_scale(tentative, modifier = 0)

adjust_kappa(tentative, modifier = 0)
}
\arguments{
\item{tentative}{\code{[numeric]} \cr The tentative parameter estimate.}

\item{modifier}{\code{[numeric=0]} \cr The modifier to adjust the tentative estimate.}
}
\description{
Functions aiding parameter adjustment after fitting an integrated step selection function (iSSF).
}
