% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_parameters.R
\name{adjust_params}
\alias{adjust_params}
\title{Adjust Movement Parameters.}
\usage{
adjust_params(x)
}
\arguments{
\item{x}{A fitted conditional logistic regression}
}
\description{
Adjusting movement based parameters, based on Avgar et al. 2016.
}
