% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amt.R
\docType{package}
\name{amt-package}
\alias{amt}
\alias{amt-package}
\title{amt: Animal Movement Tools}
\description{
Manage and analyze animal movement data. The functionality of 'amt' includes methods to calculate home ranges, track statistics (e.g. step lengths, speed, or turning angles), prepare data for fitting habitat selection analyses, and simulation of space-use from fitted step-selection functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmsigner/amt}
}

}
\author{
\strong{Maintainer}: Johannes Signer \email{jsigner@gwdg.de}

Other contributors:
\itemize{
  \item Brian Smith [contributor]
  \item Bjoern Reineking [contributor]
  \item Ulrike Schlaegel [contributor]
  \item John Fieberg [contributor]
  \item Josh O'Brien [contributor]
  \item Bernardo Niebuhr [contributor]
  \item Alec Robitaille [contributor]
  \item Scott LaPoint [data contributor]
}

}
