% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fisher.R
\docType{data}
\name{amt_fisher_covar}
\alias{amt_fisher_covar}
\title{Environmental data for fishers}
\format{
A list with three where each entry is a \code{RasterLayer}.
}
\source{
\url{https://lpdaac.usgs.gov/dataset_discovery/aster/aster_products_table/astgtm_v002}

\url{http://dup.esrin.esa.it/page_globcover.php]

\url{http://sedac.ciesin.columbia.edu/data/collection/gpw-v3/sets/browse}
}
\usage{
amt_fisher_covar
}
\description{
A list with three entries that correspond to the following three layer:
\itemize{
\item land use
\item elevation
\item population density.
}
}
\keyword{datasets}
