% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fisher.R
\docType{data}
\name{amt_fisher_lu}
\alias{amt_fisher_lu}
\title{Landuse for fisher data}
\format{
A \code{RasterLayer}
}
\usage{
amt_fisher_lu
}
\description{
A Gauss-Random-Field simulation of a landscape.
}
\keyword{datasets}
