% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animal_sim.R
\docType{data}
\name{animal_sim}
\alias{animal_sim}
\alias{animal_sim}
\title{Simulated movement data of 1 animal.}
\format{A \code{data_frame} with 32400 rows and 3 variables:
\describe{
\item{x_}{the x-coordinate}
\item{y_}{the y-coordinate}
\item{t_}{the timestamp}
}}
\usage{
animal_sim

animal_sim
}
\description{
32400 simulated relocations of an animal.

A Gauss-Random-Field simulation of a landscape.
}
\keyword{datasets}
