% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rss.R, R/log_rss_old.R
\name{append_x1}
\alias{append_x1}
\title{Append "_x1"}
\usage{
append_x1(string)

append_x1(string)
}
\arguments{
\item{string}{\verb{[character]} Variable name to possibly append to}
}
\value{
A string.

A string.
}
\description{
Helper function to append \verb{_x1} to variable names

Helper function to append \verb{_x1} to variable names
}
\details{
The function first checks if "_x1" is already appended and adds it if
it is not. This is meant for internal use in \verb{\link{plot.log_rss}()}.

The function first checks if "_x1" is already appended and adds it if
it is not. This is meant for internal use in \verb{\link{plot.log_rss}()}.
}
\keyword{internal}
