% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{as.data.frame.uhc_data}
\alias{as.data.frame.uhc_data}
\title{Coerce a \code{uhc_data} object to \code{data.frame}}
\usage{
\method{as.data.frame}{uhc_data}(uhc_data)
}
\arguments{
\item{uhc_data}{\verb{[uhc_data]} An object of class \code{uch_data}, as returned
by the function \code{\link{prep_uhc}()}.}
}
\description{
Coerces \code{uhc_data} from \code{list} to \code{data.frame}
}
