% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_sf_lines}
\alias{as_sf_lines}
\title{Export track to lines}
\usage{
as_sf_lines(x, ...)
}
\arguments{
\item{x}{\code{[track_xy, track_xyt]} \cr A track created with \code{make_track}.}

\item{...}{Further arguments, none implemented.}
}
\value{
A \code{tibble} with a \code{sfc}-column
}
\description{
Exports a track to (multi)lines from the \code{sf} package.
}
