% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{as_sf_points}
\alias{as_sf_points}
\title{Coerces a track to points}
\usage{
as_sf_points(x, ...)
}
\arguments{
\item{x}{\code{[track_xy, track_xyt]} \cr A track created with \code{make_track}.}

\item{...}{Further arguments, none implemented.}
}
\value{
A data \code{data.frame} with a \code{sfc}-column
}
\description{
Coerces a track to points from the \code{sf} package.
}
