% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_kde.R
\name{bandwidth_ref}
\alias{bandwidth_ref}
\alias{hr_kde_ref}
\alias{hr_kde_ref.track_xy}
\title{Reference bandwidth}
\usage{
hr_kde_ref(x, ...)

\method{hr_kde_ref}{track_xy}(x, rescale = "none", ...)
}
\arguments{
\item{x}{\code{[track_xy, track_xyt]} \cr A track created with \code{make_track}.}

\item{...}{Further arguments, none implemented.}

\item{rescale}{\verb{[character(1)]} \cr Rescaling method for reference bandwidth calculation. Must be one of "unitvar", "xvar", or "none".}
}
\value{
The estimated bandwidth in x and y direction.
}
\description{
Calculate the reference bandwidth for kernel density home-range range estimates.
}
