% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{bbox}
\alias{bbox}
\title{Get bounding box of a track}
\usage{
bbox(x, ...)
}
\arguments{
\item{x}{A track.}

\item{spatial}{Logical, whether or not to return a \code{sp}-object.}

\item{buffer}{Numeric, an optional buffer.}
}
\description{
Get bounding box of a track
}
