% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rss.R
\name{bootstrap_logrss}
\alias{bootstrap_logrss}
\alias{bootstrap_logrss.fit_logit}
\alias{bootstrap_logrss.fit_clogit}
\title{Bootstrap log-RSS estimate}
\usage{
bootstrap_logrss(object, ...)

\method{bootstrap_logrss}{fit_logit}(object, x1, x2, ci_level, n_boot, mle)

\method{bootstrap_logrss}{fit_clogit}(object, x1, x2, ci_level, n_boot, mle)
}
\description{
Use empirical bootstrap to estimate log-RSS CI
}
\details{
This function is meant for internal use by \code{log_rss()} and is
not meant to be called by the user.
}
\keyword{internal}
