% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{calc_w}
\alias{calc_w}
\title{Calculate \code{w(x)}}
\usage{
calc_w(f, b, newdata)
}
\arguments{
\item{f}{\verb{[formula]} \cr A model formula.}

\item{b}{\verb{[numeric]} A named vector of coefficients.}

\item{newdata}{\verb{[data.frame]} \cr A \code{data.frame} to predict eHSF values.}
}
\description{
Calculates the value of the exponential habitat selection function
}
\details{
This is actually like to be w(x) * \phi(x) for an iSSF.
}
