% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_capture_effect.R
\name{capture_effect}
\alias{capture_effect}
\alias{rm_start}
\alias{rm_end}
\alias{rm_both}
\title{Check for Capture Effects.}
\usage{
rm_start(x, time_span = lubridate::days(1))

rm_end(x, time_span = lubridate::days(1))

rm_both(x, time_span = lubridate::days(1))
}
\arguments{
\item{x}{A \code{data_table} .}

\item{capture_effect_duration}{A an object of class period, indicating the length of putative capture effects.}
}
\description{
Check for Capture Effects.
}
