% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{check_trackXY}
\alias{check_trackXY}
\alias{check_trackXYZ}
\alias{check_trackXYT}
\alias{check_trackXYZT}
\title{Check if tibble is a track}
\usage{
check_trackXY(x)

check_trackXYZ(x)

check_trackXYT(x)

check_trackXYZT(x)
}
\arguments{
\item{x}{A tibble.}
}
\value{
\code{TRUE} if tracks is as expected, otherwise an error is thrown.
}
\description{
Returns \code{TRUE} if a tibble can safely be considered as a \code{track*} and error otherwise.
}
