% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convert_angles}
\alias{convert_angles}
\alias{as_rad}
\alias{as_degree}
\title{Converts angles to radians}
\usage{
as_rad(x)

as_degree(x)
}
\arguments{
\item{x}{\verb{[numeric]}\cr Angles in degrees or rad.}
}
\description{
Converts angles to radians
}
\examples{
as_rad(seq(-180, 180, 30))

# The default unit of turning angles is rad.
data(deer)
deer \%>\% steps() \%>\% mutate(ta_ = as_degree(ta_))
}
