% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{coords}
\alias{coords}
\title{Coordinates of a track.}
\usage{
coords(x, ...)
}
\arguments{
\item{x}{\code{[track_xy, track_xyt]} \cr A track created with \code{make_track}.}

\item{...}{Further arguments, none implemented.}
}
\value{
\code{[data_frame]} \cr The coordinates.
}
\description{
Coordinates of a track.
}
\examples{
data(deer)
coords(deer)
}
