% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{crs}
\alias{crs}
\alias{get_crs}
\alias{has_crs}
\title{Coordinate References System (CRS)}
\usage{
get_crs(x, ...)

has_crs(x, ...)
}
\arguments{
\item{x}{\code{[any]} \cr Object to check.}

\item{...}{Further arguments, none implemented.}
}
\description{
Check if an object has a coordinate reference system (\code{has_crs}) and returns the \code{proj4string} with \code{get_crs} of the coordinate reference system.
}
\examples{
data(deer)
has_crs(deer)
get_crs(deer)
}
