% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_ud.R
\name{cum_ud}
\alias{cum_ud}
\alias{cumulative_ud}
\alias{cumulative_ud.RasterLayer}
\title{Calculate a cumulative UD}
\usage{
cumulative_ud(x, ...)

\method{cumulative_ud}{RasterLayer}(x, ...)
}
\arguments{
\item{x}{\code{[RasterLayer]} \cr Containig the Utilization Distribution (UD).}

\item{...}{Further arguments, none implemented.}
}
\value{
\code{[RasterLayer]} \cr The cumulative UD.
}
\description{
Calculate a cumulative UD
}
\note{
This function is typically used to obtain isopleths.
}
