% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_ud.R
\name{cumulative_ud}
\alias{cumulative_ud}
\title{Calculate a cumulative UD}
\usage{
cumulative_ud(x, ...)
}
\arguments{
\item{x}{A UD.}

\item{...}{Further arguments, none implemented.}
}
\description{
Calculate a cumulative UD
}
