% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_night.R
\name{day_night}
\alias{day_night}
\alias{day_night.steps}
\title{Extract if a fix was taken during day or night.}
\usage{
day_night(x, ...)

\method{day_night}{steps}(x, end = TRUE, ...)
}
\arguments{
\item{x}{A \code{track_xyt} or \code{steps_xyt}}

\item{...}{Further arguments, none implemented.}

\item{end}{A logical scalar, indicating if the time stemp of the end of the step should be used.}
}
\description{
A convinience wrapper around \code{maptools::sunriset}.
}
