% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sh.R
\docType{data}
\name{deer}
\alias{deer}
\title{Relocations of 1 red deer.}
\format{A track_xyt
\describe{
\item{x_epsg31467}{the x-coordinate}
\item{y_epsg31467}{the y-coordinate}
\item{t_}{time stamp}
}}
\source{
Verein für Wildtierforschung Dresden und Göttingen e.V.
}
\usage{
deer
}
\description{
1500 GPS relocations of one red deer from northern Germany.
}
\keyword{datasets}
