% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{diff}
\alias{diff}
\alias{diff_x}
\alias{diff_y}
\title{Difference in x and y}
\usage{
diff_x(x, ...)

diff_y(x, ...)
}
\arguments{
\item{x}{A track_xy{t}.}

\item{...}{Further arguments, none implemented.}
}
\value{
Numeric vector
}
\description{
Difference in x and y coordinates.
}
