% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_support.R
\name{diff_matrix}
\alias{diff_matrix}
\alias{diff_matrix.glm}
\alias{diff_matrix.clogit}
\alias{diff_matrix.gam}
\title{Return difference matrix}
\usage{
diff_matrix(object, x1, x2, ...)

\method{diff_matrix}{glm}(object, x1, x2, ...)

\method{diff_matrix}{clogit}(object, x1, x2, ...)

\method{diff_matrix}{gam}(object, x1, x2, ...)
}
\arguments{
\item{object}{\verb{[a fitted model]} \cr A fitted RSF or (i)SSF model. For possible
models, see \code{\link{supported_models}()}}

\item{x1}{\verb{[data.frame]} \cr A \code{data.frame} representing the habitat values
at location x_1. Must contain all fitted covariates as expected by
\code{predict()}.}

\item{x2}{\verb{[data.frame]} \cr A 1-row \code{data.frame} representing the single
hypothetical location of x_2. Must contain all fitted covariates as expected
by \code{predict()}.}

\item{...}{Further arguments, none implemented.}
}
\description{
Subtracts matrix \code{x2} from all rows of matrix \code{x1}
}
\details{
For internal use in calculating log-RSS.
}
\author{
Brian J. Smith
}
