% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_characteristics.R
\name{direction_abs}
\alias{direction_abs}
\alias{direction_abs.track_xy}
\title{Absolute Direction.
Function to calculate the absolute direction of a movement track. 0 is north.}
\usage{
direction_abs(x)

\method{direction_abs}{track_xy}(x)
}
\arguments{
\item{x}{A track.}
}
\description{
Absolute Direction.
Function to calculate the absolute direction of a movement track. 0 is north.
}
