% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_center.R
\name{dist_cent}
\alias{dist_cent}
\alias{distance_to_center}
\alias{distance_to_center.track_xy}
\alias{distance_to_centers}
\alias{distance_to_centers.track_xy}
\title{Distance to center.}
\usage{
distance_to_center(x, ...)

\method{distance_to_center}{track_xy}(x, trast, square = TRUE, ...)

distance_to_centers(x, ...)

\method{distance_to_centers}{track_xy}(x, trast, top_n = 10, square = TRUE,
  ...)
}
\arguments{
\item{x}{a \code{track_xy}.}

\item{...}{Further arguments, none implemented.}

\item{trast}{\code{[RasterLayer]} \cr A template.}

\item{square}{\code{[logical(1)]} \cr Should the distance be squared?}

\item{top_n}{\code{[integer(1)]} \cr To how many centers should the distance be calculated?}
}
\value{
\code{RasterLayer}
}
\description{
Calculates the distance of each pixel to the centroid of the track.
}
