% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_center.R
\name{distance_to_center}
\alias{distance_to_center}
\alias{distance_to_center.track_xy}
\title{Distance to center.}
\usage{
distance_to_center(x, ...)

\method{distance_to_center}{track_xy}(x, trast, ...)
}
\arguments{
\item{x}{a \code{track_xy}.}

\item{...}{Further arguments, none implemented.}

\item{trast}{A template \code{RasterLayer}.}
}
\value{
\code{RasterLayer}
}
\description{
Calculates the distance of each pixel to the centroid of the track.
}
