% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distr_name}
\alias{distr_name}
\alias{sl_distr_name}
\alias{sl_distr_name.random_steps}
\alias{sl_distr_name.fit_clogit}
\alias{ta_distr_name}
\alias{ta_distr_name.random_steps}
\alias{ta_distr_name.fit_clogit}
\title{Name of step-length distribution and turn-angle distribution}
\usage{
sl_distr_name(x, ...)

\method{sl_distr_name}{random_steps}(x, ...)

\method{sl_distr_name}{fit_clogit}(x, ...)

ta_distr_name(x, ...)

ta_distr_name(x, ...)

\method{ta_distr_name}{random_steps}(x, ...)

\method{ta_distr_name}{fit_clogit}(x, ...)
}
\arguments{
\item{x}{Random steps or fitted model}

\item{...}{None implemented.}
}
\description{
Name of step-length distribution and turn-angle distribution
}
