% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rss.R
\name{emp_boot1}
\alias{emp_boot1}
\alias{emp_boot1.clogit}
\title{Single bootstrap iteration}
\usage{
emp_boot1(object, x1, x2)

\method{emp_boot1}{clogit}(object, x1, x2)
}
\description{
Runs a single iteration of the empirical bootstrap
}
\details{
This function is meant for internal use by \code{bootstrap_logrss()} and
is not meant to be called by the user.
}
\keyword{internal}
