% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rss.R
\name{emp_bootstrap_logrss}
\alias{emp_bootstrap_logrss}
\alias{emp_bootstrap_logrss.default}
\title{Empirical bootstrap log-RSS estimate}
\usage{
emp_bootstrap_logrss(object, ...)

\method{emp_bootstrap_logrss}{default}(object, x1, x2, ci_level, n_boot, mle, ...)
}
\description{
Use empirical bootstrap to estimate log-RSS CI
}
\details{
This function is meant for internal use by \code{log_rss()} and is
not meant to be called by the user.
}
\keyword{internal}
