% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fisher.R
\docType{data}
\name{env_covar}
\alias{env_covar}
\title{Environmental data for fishers}
\format{
A list with three where each entry is a \code{RasterLayer}.
}
\source{
https://lpdaac.usgs.gov/dataset_discovery/aster/aster_products_table/astgtm_v002

http://dup.esrin.esa.it/page_globcover.php

http://sedac.ciesin.columbia.edu/data/collection/gpw-v3/sets/browse
}
\usage{
env_covar
}
\description{
A list with three entries that correspond to the following three layer:
\itemize{
\item land use
\item elevation
\item population density.
}
}
\keyword{datasets}
