% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{Extract.uhc_data}
\alias{Extract.uhc_data}
\alias{[.uhc_data}
\title{Subset a \code{uhc_data} object}
\usage{
\method{[}{uhc_data}(x, i)
}
\arguments{
\item{x}{\verb{[uhc_data]} A \code{uhc_data} object to subset.}

\item{i}{\verb{[numeric} or \verb{character]} A numeric vector to subset variables
by position or a character vector to subset variables by name.}
}
\description{
Subset a \code{uhc_data} object
}
