% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{fit_sl_dist}
\alias{fit_sl_dist}
\alias{fit_sl_dist_}
\alias{fit_sl_dist}
\alias{fit_sl_dist_base}
\title{Fit a Step-Lengths distribution.}
\usage{
fit_sl_dist_(.tbl, x, ...)

fit_sl_dist(.tbl, x, ...)

fit_sl_dist_base(x, na.rm = TRUE, distr = "gamma", ...)
}
\arguments{
\item{.tbl}{A \code{track_xy*}.}

\item{x}{The column, containing step lengths, usually \code{sl_}.}

\item{...}{Further arguments, none implemented.}

\item{na.rm}{Logical scalar, should \code{NA} be removed?}

\item{distr}{Name of the distribution, curently only \code{gamma} distribution is supported.}
}
\description{
Fit a distribution to the the step lengths of a track.
}
