% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{fit_ta_dist}
\alias{fit_ta_dist}
\alias{fit_ta_dist_}
\alias{fit_ta_dist}
\alias{fit_ta_dist_base}
\title{Fit a turning-angles distribution.}
\usage{
fit_ta_dist_(.tbl, x, ...)

fit_ta_dist(.tbl, x, ...)

fit_ta_dist_base(x, na.rm = TRUE, distr = "vonmises", ...)
}
\arguments{
\item{.tbl}{A \code{track_xy*}.}

\item{x}{The column, containing the relative turning angles, usually \code{ta_}.}

\item{na.rm}{Logical scalar, should \code{NA} be removed?}

\item{distr}{Name of the distribution, curently only \code{vonmises} distribution is supported.}
}
\description{
Fit a distribution to the the turning angles of a track.
}
