% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_to.R
\name{from_to}
\alias{from_to}
\alias{from_to.track_xyt}
\alias{from}
\alias{from.track_xyt}
\alias{to}
\alias{to.track_xyt}
\title{Duration of tracks}
\usage{
from_to(x, ...)

\method{from_to}{track_xyt}(x, ...)

from(x, ...)

\method{from}{track_xyt}(x, ...)

to(x, ...)

\method{to}{track_xyt}(x, ...)
}
\arguments{
\item{x}{\code{[track_xy, track_xyt]} \cr A track created with \code{make_track}.}

\item{...}{Further arguments, none implemented.}
}
\description{
Function that returns the start (\code{from}), end (\code{to}), and the duration (\code{from_to}) of a track.
}
\examples{
data(deer)
from(deer)
to(deer)
from_to(deer)
}
