% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fisher.R
\name{get_amt_fisher_covars}
\alias{get_amt_fisher_covars}
\title{Helper function to get fisher covars}
\usage{
get_amt_fisher_covars()
}
\value{
A list with covariates
}
\description{
The current version of \code{terra} (1.7.12) requires \code{SpatRast}ers to be wrapped in order to be saved locally. This function unwraps the covariates for the fisher data and returns a list.
}
