% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{get_crs}
\alias{get_crs}
\title{Obtains the Coordinate Reference Systems}
\usage{
get_crs(x, ...)
}
\arguments{
\item{x}{\code{[any]} \cr Object to check.}

\item{...}{Further arguments, none implemented.}
}
\value{
The \code{proj4string} of the CRS.
}
\description{
Returns the \code{proj4string} of an object.
}
\examples{
data(deer)
get_crs(deer)
}
