% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{get_displacement}
\alias{get_displacement}
\title{Calculate Expected Displacement}
\usage{
get_displacement(delta, time_span)
}
\arguments{
\item{delta}{\verb{[numeric]} A squared displacement rate (SDR), such as that
returned by \code{\link{calculate_sdr}()}. Units assumed to be \code{m^2/s}.}

\item{time_span}{\verb{[Period]} A \code{lubridate} \code{Period} giving the time span
for which to calculate expected displacement.}
}
\value{
Returns a numeric vector (of length 1) with the expected displacement
in meters.
}
\description{
Calculates expected displacement for a given SDR and time span
}
\seealso{
\code{\link{calculate_sdr}()}
}
\author{
Johannes Signer and Brian J. Smith
}
