% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{get_distr}
\alias{get_distr}
\alias{sl_distr}
\alias{sl_distr.random_steps}
\alias{sl_distr.fit_clogit}
\alias{ta_distr}
\alias{ta_distr.random_steps}
\alias{ta_distr.fit_clogit}
\title{Get distribution and parameters}
\usage{
sl_distr(x, ...)

\method{sl_distr}{random_steps}(x, ...)

\method{sl_distr}{fit_clogit}(x, ...)

ta_distr(x, ...)

\method{ta_distr}{random_steps}(x, ...)

\method{ta_distr}{fit_clogit}(x, ...)
}
\arguments{
\item{x}{Random steps or fitted model}

\item{...}{None implemented.}
}
\description{
Get distribution and parameters
}
