% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_max_dist}
\alias{get_max_dist}
\alias{get_max_dist.fit_clogit}
\title{Get the maximum distance}
\usage{
get_max_dist(x, ...)

\method{get_max_dist}{fit_clogit}(x, p = 0.99, ...)
}
\arguments{
\item{x}{\verb{[fitted_issf]} \cr A fitted integrated step-selection function.}

\item{...}{Further arguments, none implemented.}

\item{p}{\verb{[numeric(1)]\{0.99\}} The quantile of the step-length distribution.}
}
\description{
Helper function to get the maximum distance from a fitted model.
}
