% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sh.R
\name{get_sh_forest}
\alias{get_sh_forest}
\title{Helper function to get forest cover}
\usage{
get_sh_forest()
}
\value{
A \code{SpatRast} with forest cover.
}
\description{
The current version of \code{terra} (1.7.12) requires \code{SpatRast}ers to be wrapped in order to be saved locally. This function unwraps the the forest layer and returns a \code{SpatRast}.
}
