% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_support.R
\name{get_Sigma}
\alias{get_Sigma}
\alias{get_Sigma.default}
\title{Return variance-covariance matrix}
\usage{
get_Sigma(object, ...)

\method{get_Sigma}{default}(object, ...)
}
\arguments{
\item{object}{\verb{[a fitted model]} \cr A fitted RSF or (i)SSF model. For possible
models, see \code{\link{supported_models}()}}

\item{...}{Further arguments, none implemented.}
}
\description{
Returns a variance-covariance matrix for any supported model type
}
\details{
The \code{default} method is confirmed to work for:
\itemize{
\item \code{glm}
\item \code{clogit}
\item \code{gam}
}
}
