% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-animal_sim.R
\docType{data}
\name{habitat}
\alias{habitat}
\title{Habitat of simulated data}
\format{A \code{RasterLayer}}
\usage{
habitat
}
\description{
A Gauss-Random-Field simulation of a landscape.
}
\keyword{datasets}
