% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{has_crs}
\alias{has_crs}
\title{Check for Coordinate Reference Systems (CRS)}
\usage{
has_crs(x, ...)
}
\arguments{
\item{x}{\code{[any]} \cr Object to check.}

\item{...}{Further arguments, none implemented.}
}
\value{
Logic vector of length 1.
}
\description{
Checks if an object has a CRS.
}
\examples{
data(deer)
has_crs(deer)
}
