% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_area.R
\name{hr_area}
\alias{hr_area}
\alias{hr_area.hr}
\alias{hr_area.SpatRaster}
\alias{hr_area.akde}
\title{Home-range area}
\usage{
hr_area(x, ...)

\method{hr_area}{hr}(x, units = FALSE, ...)

\method{hr_area}{SpatRaster}(x, level = 0.95, ...)

\method{hr_area}{akde}(x, units = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{hr}}

\item{...}{Further arguments, none implemented.}

\item{units}{\verb{[logic(1)]} \cr Should areas be returned as units? If \code{FALSE} areas are returned as numeric values.}

\item{level}{The level at which the area will be calculated.}
}
\value{
A \code{tibble} with the home-range level and the area.
}
\description{
Obtain the area of a home-range estimate, possible at different isopleth levels.
}
