% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_overlap.R
\name{hr_overlap_feature}
\alias{hr_overlap_feature}
\title{Calculate the overlap between a home-range estimate and a polygon}
\usage{
hr_overlap_feature(x, sf, direction = "hr_with_feature", feature_names = NULL)
}
\arguments{
\item{x}{A home-range estimate.}

\item{sf}{An object of class \code{sf} containing polygons}

\item{direction}{The direction.}

\item{feature_names}{optional feature names}
}
\value{
A tibble
}
\description{
Sometimes the percentage overlap between a spatial polygon an a home-range is required.
}
