% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insepct.R
\name{inspect}
\alias{inspect}
\alias{inspect.track_xy}
\title{Inspect a track}
\usage{
inspect(x, ...)

\method{inspect}{track_xy}(x, popup = NULL, cluster = TRUE, ...)
}
\arguments{
\item{x}{A \code{track_xy}. Important, \code{x} requires a valid coordinate reference system.}

\item{...}{Further arguments, none implemented.}

\item{popup}{A character vector of length == \code{nrow(x)}, optionally providing popups.}

\item{cluster}{A locigcal scalar, indicating if points should be clustered at lower zoom levels.}
}
\description{
Provides a very basic interfact to \code{leaflet} and lets the user inspect relocations on an interactive map.
}
\examples{
data(sh)
x <- track(x = sh$x, y = sh$y, crs = sp::CRS("+init=epsg:31467"))

\dontrun{
inspect(x)
inspect(x, cluster = FALSE)
inspect(x, popup = 1:nrow(x), cluster = FALSE)
}
}
\seealso{
\code{leaflet::leaflet()}
}
