% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{is_trackXY}
\alias{is_trackXY}
\alias{is_trackXYZ}
\alias{is_trackXYT}
\alias{is_trackXYZT}
\title{Check if tibble is a track}
\usage{
is_trackXY(x)

is_trackXYZ(x)

is_trackXYT(x)

is_trackXYZT(x)
}
\arguments{
\item{x}{A tibble.}
}
\value{
\code{TRUE} if tracks is as expected, otherwise \code{FALSE}.
}
\description{
Returns \code{TRUE} if a tibble can safely be considered as a \code{track*}.
}
