% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{issf_drop_stratum}
\alias{issf_drop_stratum}
\title{Create formula without stratum from iSSF}
\usage{
issf_drop_stratum(object, l)
}
\arguments{
\item{object}{\verb{[fit_clogit]} Fitted iSSF.}

\item{l}{\verb{[list]} List returned by \code{prep_test_dat.fit_clogit()}}
}
\description{
Creates a formula without stratum variable
}
