% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{issf_w_form}
\alias{issf_w_form}
\title{Create habitat formula from iSSF}
\usage{
issf_w_form(object, l)
}
\arguments{
\item{object}{\verb{[fit_clogit]} Fitted iSSF.}

\item{l}{\verb{[list]} List returned by \code{prep_test_dat.fit_clogit()}}
}
\description{
Creates a formula without movement variables
}
