% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_kernel.R
\name{kernel_add_covars}
\alias{kernel_add_covars}
\title{Internal function to annotate a dispersal kernel with (spatial) covariates}
\usage{
kernel_add_covars(kernel, spatial.covars, position, temporal.covars = NULL)
}
\arguments{
\item{kernel}{Ther dispersal kernel}

\item{spatial.covars}{RasterStack with covariates}

\item{position}{The current position.}

\item{temporal.covars}{Additional covariates (non sptatially varying).}
}
\description{
Internal function to annotate a dispersal kernel with (spatial) covariates
}
