% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_kernel.R
\name{kernel_finish}
\alias{kernel_finish}
\title{Internal function to "finish" a dispersal kernel.}
\usage{
kernel_finish(
  kernel,
  formula,
  coefficients,
  return.raster = FALSE,
  normalize = TRUE
)
}
\arguments{
\item{kernel}{The dispersal kernel.}

\item{formula}{The definition of the kernel from the covariates.}

\item{coefficients}{The values of the coefficients (need to match formula).}

\item{return.raster}{Should a raster be returned or not?}

\item{normalize}{Should the kernel be normaliszed}
}
\description{
Internal function to "finish" a dispersal kernel.
}
