% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_kernel.R
\name{kernel_rotate}
\alias{kernel_rotate}
\title{Internal function to rotate a movement kernel}
\usage{
kernel_rotate(kernel, direction)
}
\arguments{
\item{kernel}{The movement kernel.}

\item{direction}{The direction of the rotation. Clockwise in degrees.}
}
\description{
Internal function to rotate a movement kernel
}
