% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_kernel.R
\name{kernel_setup}
\alias{kernel_setup}
\title{Internal function to create a movement kernel}
\usage{
kernel_setup(template, max.dist = 100, position = c(0, 0))
}
\arguments{
\item{template}{A raster (stack) that is the template for the simulations.}

\item{max.dist}{The maximum distance of the dispersal kernel}

\item{position}{The center of the dispersal kernel.}
}
\description{
Internal function to create a movement kernel
}
