% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_kernel.R
\name{kernel_shift}
\alias{kernel_shift}
\title{Internal function to shift a movement kernel}
\usage{
kernel_shift(kernel, dxy)
}
\arguments{
\item{kernel}{The movement kernel.}

\item{dxy}{The amount of x and y to shift ther kernel.}
}
\description{
Internal function to shift a movement kernel
}
