% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_support.R
\name{linear_predictor}
\alias{linear_predictor}
\alias{linear_predictor.glm}
\alias{linear_predictor.clogit}
\alias{linear_predictor.gam}
\title{Return linear predictor}
\usage{
linear_predictor(object, newdata, ...)

\method{linear_predictor}{glm}(object, newdata, ...)

\method{linear_predictor}{clogit}(object, newdata, ...)

\method{linear_predictor}{gam}(object, newdata, ...)
}
\arguments{
\item{object}{\verb{[a fitted model]} \cr A fitted RSF or (i)SSF model. For possible
models, see \code{\link{supported_models}()}}

\item{newdata}{\verb{[data.frame]} \cr A \code{data.frame}. Must contain all fitted
covariates as expected by \code{predict()}.}

\item{...}{Further arguments, none implemented.}
}
\description{
Returns the linear predictor for any supported model type
}
\details{
For internal use in calculating log-RSS.
}
\author{
Brian J. Smith
}
