% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_locoh.R
\name{locoh_k}
\alias{locoh_k}
\alias{locoh_k.track_xy}
\title{Local Convex Hull}
\usage{
locoh_k(x, ...)

\method{locoh_k}{track_xy}(x, n = 10, level = 0.95, rand_buffer = 1e-05,
  ...)
}
\arguments{
\item{x}{A \code{track_xy}.}

\item{n}{Numeric scalar, the number of neighbours.}

\item{level}{Numeric scalar/vector, home-range level. The level should be \code{0 < level < 1}.}

\item{rand_buffer}{Numeric scalar, random buffer to avaoid polygons with area 0 (if coordinates are numerically identical).}
}
\description{
Calculate Local Convex Hull (LoCoH) home range type "k". That is, the k nearest neighbours are considered.
}
\examples{
data(sh)
x <- track(x = sh[, 1], y = sh[, 2])
l <- locoh_k(x)
}
