% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_issf_model}
\alias{make_issf_model}
\title{Create an \code{issf}-model object from scratch}
\usage{
make_issf_model(
  coefs = c(sl_ = 0),
  sl = make_exp_distr(),
  ta = make_unif_distr()
)
}
\arguments{
\item{coefs}{A named vector with the coefficient values.}

\item{sl}{The tentative step-length distribution.}

\item{ta}{The tentative turn-angle distribution.}
}
\value{
An object of \code{fit_clogit}.
}
\description{
In order to simulate from an \code{issf} a
}
