% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement2.R
\name{max_dist}
\alias{max_dist}
\title{Calculate a maximum distance from a \code{steps_xy*} object}
\usage{
max_dist(obj, fctr = 1.5)
}
\arguments{
\item{obj}{\code{[steps_xyt, steps_xy]} A \code{steps_xy*} object from which to extract maximum observed step length}

\item{fctr}{\code{[numeric(1) = 1.5]} Factor by which to increase maximum distance. See details}
}
\value{
\code{numeric} vector of length 1
}
\description{
Calculate a maximum distance from a \code{steps_xy*} object
}
\details{
This function takes a dataset of class \code{steps_xy*} and returns a
sensible value for the \code{max.dist} argument of \code{\link{dispersal_kernel}()}. The
result is given by \code{max(obj$sl_) * fctr}.
}
