% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_support.R
\name{model_matrix}
\alias{model_matrix}
\alias{model_matrix.glm}
\alias{model_matrix.clogit}
\alias{model_matrix.gam}
\title{Return model matrix}
\usage{
model_matrix(object, newdata, ...)

\method{model_matrix}{glm}(object, newdata, ...)

\method{model_matrix}{clogit}(object, newdata, ...)

\method{model_matrix}{gam}(object, newdata, ...)
}
\arguments{
\item{object}{\verb{[a fitted model]} \cr A fitted RSF or (i)SSF model. For possible
models, see \code{\link{supported_models}()}}

\item{newdata}{\verb{[data.frame]} \cr A \code{data.frame}. Must contain all fitted
covariates as expected by \code{predict()}.}

\item{...}{Further arguments, none implemented.}
}
\description{
Returns model matrix for \code{newdata}
}
\details{
For internal use in calculating log-RSS.
}
\author{
Brian J. Smith
}
