% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement_metrics.R
\name{movement_metrics}
\alias{movement_metrics}
\alias{straightness}
\alias{cum_dist}
\alias{tot_dist}
\alias{msd}
\alias{intensity_use}
\alias{sinuosity}
\alias{tac}
\title{Different metrics for a movement track}
\usage{
straightness(x, ...)

cum_dist(x, ...)

tot_dist(x, ...)

msd(x, ...)

intensity_use(x, ...)

sinuosity(x, ...)

tac(x, ...)
}
\arguments{
\item{x}{A \code{track_xy{y}}}

\item{...}{Further arguments, none implemented.}
}
\description{
Functions to calculate metrics such as straightness, mean squred displacement (msd), intensity use, for a track.
}
\details{
The intensity use is calculated by dividing the total movement distance by the square of the area of movement (= minimum convex polygon 100),
sinuosity, mean turn angle correlation (tac) .
}
\examples{
data("sh")
sh <- mk_track(sh, x_epsg31467, y_epsg31467)

tot_dist(sh)
cum_dist(sh)
straightness(sh)
msd(sh)
intensity_use(sh)

}
\references{
\insertRef{almeida2010}{amt}
\insertRef{swihart1985}{amt}
}
