% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsd.R
\name{nsd}
\alias{nsd}
\title{Net squared displacement (nsd)}
\usage{
nsd(x, ...)
}
\arguments{
\item{x}{A track_xy{t}.}

\item{...}{Further arguments, none implemented.}
}
\value{
Numeric vector.
}
\description{
Calculate the net squared displacement (i.e., the squared distance from the first location of a track) for a track.
}
