% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rss.R
\name{para_boot1}
\alias{para_boot1}
\alias{para_boot1.default}
\title{Single bootstrap iteration}
\usage{
para_boot1(object, x1, x2)

\method{para_boot1}{default}(object, x1, x2)
}
\description{
Runs a single iteration of the parametric bootstrap
}
\details{
This function is meant for internal use by \code{bootstrap_logrss()} and
is not meant to be called by the user.
}
\keyword{internal}
