% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rss.R
\name{para_bootstrap_logrss}
\alias{para_bootstrap_logrss}
\alias{para_bootstrap_logrss.default}
\title{Parametric bootstrap log-RSS estimate}
\usage{
para_bootstrap_logrss(object, ...)

\method{para_bootstrap_logrss}{default}(object, x1, x2, ci_level, n_boot, ...)
}
\description{
Use parametric bootstrap to estimate log-RSS CI
}
\details{
This function is meant for internal use by \code{log_rss()} and is
not meant to be called by the user.
}
\keyword{internal}
