% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{params}
\alias{params}
\alias{sl_distr_params}
\alias{sl_distr_params.random_steps}
\alias{sl_distr_params.fit_clogit}
\alias{ta_distr_params}
\alias{ta_distr_params.random_steps}
\alias{ta_distr_params.fit_clogit}
\title{Get parameters from a (fitted) distribution}
\usage{
sl_distr_params(x, ...)

\method{sl_distr_params}{random_steps}(x, ...)

\method{sl_distr_params}{fit_clogit}(x, ...)

ta_distr_params(x, ...)

\method{ta_distr_params}{random_steps}(x, ...)

\method{ta_distr_params}{fit_clogit}(x, ...)
}
\arguments{
\item{x}{\verb{[amt_distr]}\cr A (fitted) distribution}

\item{...}{None}
}
\value{
A list with the parameters of the distribution.
}
\description{
Get parameters from a (fitted) distribution
}
\examples{
data(deer)
d <- deer \%>\% steps() \%>\% random_steps()
sl_distr_params(d)
ta_distr_params(d)


}
