% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_plot.R
\name{plot.hr}
\alias{plot.hr}
\title{Plot a home-range estimate}
\usage{
\method{plot}{hr}(x, add.relocations = TRUE, ...)
}
\arguments{
\item{x}{A home-range estimate.}

\item{add.relocations}{\code{logical(1)} indicates if a relocations should be
added to the plot.}

\item{...}{Further arguments, none implemented.}
}
\value{
A plot
}
\description{
Plot a home-range estimate
}
