% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{plot.uhc_data}
\alias{plot.uhc_data}
\title{Plot UHC plots}
\usage{
\method{plot}{uhc_data}(uhc_data)
}
\arguments{
\item{uhc_data}{\verb{[uhc_data]} An object of class \code{uch_data}, as returned
by the function \code{\link{prep_uhc}()}.}
}
\description{
Plot an object of class \code{uhc_data}
}
\details{
Makes plots mimicking those in Fieberg et al. (2018), with the
bootstrapped distribution in gray, the observed distribution in black,
and the available distribution as a dashed red line.
}
\seealso{
\code{\link{prep_uhc}()}
}
