% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{plot.uhc_envelopes}
\alias{plot.uhc_envelopes}
\title{Plot simplified UHC plots}
\usage{
\method{plot}{uhc_envelopes}(x, ...)
}
\arguments{
\item{x}{\verb{[uhc_envelopes]} An object of class \code{uhc_envelopes}, as returned
by the function \code{\link{conf_envelope}()}.}

\item{...}{Included for consistency with generic
\code{\link{plot}()}. Currently ignored.}
}
\description{
Plot an object of class \code{uhc_envelopes}
}
\details{
Makes plots mimicking those in Fieberg et al. (2018), with the
bootstrapped distribution in gray, the observed distribution in black,
and the available distribution as a dashed red line. This differs from
\code{\link{plot.uhc_data}()} in that the bootstrapped distribution
(in gray) is drawn as a polygon rather than (many) lines, speeding up
plotting performance.
}
\seealso{
\code{\link{prep_uhc}()}, \code{\link{conf_envelope}()},
\code{\link{plot.uhc_data}()}
}
\author{
Brian J. Smith
}
