% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{plot_sl}
\alias{plot_sl}
\title{Plot step-length distribution}
\usage{
plot_sl(x, ...)
}
\arguments{
\item{x}{\code{[fit_clogit]} \cr A fitted step selection.}

\item{...}{Further arguments, none implemented.}
}
\description{
Plot step-length distribution
}
