% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv.R
\name{prep_mgcv}
\alias{prep_mgcv}
\title{Format data for \code{mgcv}}
\usage{
prep_mgcv(object, ...)
}
\arguments{
\item{object}{\verb{[data.frame]} \cr The formatted data. Often an object of
class \code{random_steps} (which is ultimately a \code{data.frame}).}

\item{...}{Further arguments, none implemented.}
}
\value{
Returns data with same class as input.
}
\description{
Adds columns to data for use in (i)SSF fitting via \code{mgcv::gam()}
}
\details{
This function currently only adds a single column of 1s to the
data, labeled \code{times_} following Klappstein et al. (2024).
}
\examples{

\donttest{

# Load packages
library(mgcv)

# Load data
data("deer")
sh_forest <- get_sh_forest()

# Prepare data for SSF
ssf_data <- deer |>
  steps_by_burst() |>
  random_steps(n = 15) |>
  extract_covariates(sh_forest) |>
  mutate(forest = factor(forest, levels = 1:0,
                         labels = c("forest", "non-forest"))) |>
  # Prepare data for mgcv
  prep_mgcv()

}
}
\references{
Klappstein, N.J., Michelot, T., Fieberg, J., Pedersen, E.J. & Mills-Flemming, J. (2024). Step selection functions with non‐linear and random effects. \emph{Methods in Ecology and Evolution}, 15(8), 1332-1346. doi: 10.1111/2041-210X.14367
}
\author{
Brian J. Smith
}
