% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{prep_test_dat}
\alias{prep_test_dat}
\title{Prepares \code{test_dat} for \code{uhc_prep()}}
\usage{
prep_test_dat(object, test_dat, verbose = TRUE)
}
\arguments{
\item{object}{\verb{[glm, fit_logit, fit_clogit]} \cr A fitted RSF or (i)SSF model.
Should be fit to \emph{training} dataset separate from the testing data.}

\item{test_dat}{\verb{[data.frame]} \cr A \code{data.frame} with \emph{testing} data from
which to sample test points. Should be separate from the data used to train
the model passed to \code{object}.}

\item{verbose}{\verb{[logical]} Should messages be displayed (\code{TRUE}) or not
(\code{FALSE})?}
}
\description{
Internal function to check and format \code{test_dat}
}
\keyword{internal}
