% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{random_numbers}
\alias{random_numbers}
\title{Sample random numbers}
\usage{
random_numbers(x, n = 100, ...)
}
\arguments{
\item{x}{\verb{[amt_distr]} \cr A distribution object.}

\item{n}{\verb{[integer(1)=100]\{>0\}} \cr The number of random draws.}

\item{...}{none implemented.}
}
\value{
A numermic vector.
}
\description{
Sample radom numbers from a distribution created within the \code{amt} package.
}
