% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_points.R
\name{random_points}
\alias{random_points}
\alias{random_points.mcp}
\alias{random_points.track_xy}
\title{Generate Random Points.}
\usage{
random_points(x, ...)

\method{random_points}{mcp}(x, n = 100, type = "random", ...)

\method{random_points}{track_xy}(x, level = 1, hr = "mcp", factor = 10,
  ...)
}
\arguments{
\item{x}{A track, or a home range.}

\item{...}{Further arguments, none implemented.}

\item{n}{The number of random points.}

\item{type}{Argument passed to \code{sp::spsample type}.}

\item{level}{Home range level.}

\item{hr}{The home range estimator to be used.}

\item{factor}{How many random points shoud be estimated for each observed point?}
}
\description{
Functions to generate random points within an animals home range. This is usually used for Resource Selection Functions (RSF).
}
