% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{random_steps_simple}
\alias{random_steps_simple}
\title{Simulate from an ssf model}
\usage{
random_steps_simple(start, sl_model, ta_model, n.control)
}
\arguments{
\item{start}{First step}

\item{sl_model}{Step length model to use}

\item{ta_model}{Turning angle model to use}

\item{n.control}{How many alternative steps are considered each step}
}
\value{
Simulated trajectory
}
\description{
Simulate from an ssf model
}
