% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\alias{range_x}
\alias{range_x.track_xy}
\alias{range_y}
\alias{range_y.track_xy}
\alias{range_both}
\alias{range_both.track_xy}
\title{Geographic range}
\usage{
range_x(x, ...)

\method{range_x}{track_xy}(x, ...)

range_y(x, ...)

\method{range_y}{track_xy}(x, ...)

range_both(x, ...)

\method{range_both}{track_xy}(x, ...)
}
\arguments{
\item{x}{\code{[track_xy, track_xyt]} \cr A track created with \code{make_track}.}

\item{...}{Further arguments, none implemented.}
}
\value{
Numeric vector with the range.
}
\description{
The range that in either \code{x}, \code{y} or \code{both} directions, that a track covers.
}
