% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement2.R
\name{raster_center}
\alias{raster_center}
\title{Find the center of a raster}
\usage{
raster_center(r)
}
\arguments{
\item{r}{\code{[Raster*]} A \code{RasterLayer} or similar which has an \code{Extent} object}
}
\description{
Find the center of a raster
}
