% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMB.R
\name{RE_glmmTMB}
\alias{RE_glmmTMB}
\title{Return names of random effects in glmmTMB model}
\usage{
RE_glmmTMB(object)
}
\arguments{
\item{object}{\verb{[glmmTMB]} \cr A \code{glmmTMB} model.}
}
\description{
Internal function to return random effects and identify strata variable
}
