% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_capture_effect.R
\name{remove_capture}
\alias{remove_capture}
\alias{remove_capture_effect}
\alias{remove_capture_effect.track_xyt}
\title{Removes Capture Effects}
\usage{
remove_capture_effect(x, ...)

\method{remove_capture_effect}{track_xyt}(x, start, end, ...)
}
\arguments{
\item{x}{An object of class \code{track_xyt}.}

\item{...}{Further arguments, none implemented.}

\item{start}{A \code{lubirdate::Period}, indicating the time period to be removed at the beginning of the track.}

\item{end}{A \code{lubirdate::Period}, indicating the time period to be removed at the end of the track.}
}
\value{
A \code{tibble} without observations that fall within the period of the capture effect.
}
\description{
Removing relocations at the beginning and/or end of a track, that fall within a user specified period.
}
