% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_support.R
\name{resample_betas}
\alias{resample_betas}
\alias{resample_betas.default}
\alias{resample_betas.fit_clogit}
\title{Resample betas}
\usage{
resample_betas(object, ...)

\method{resample_betas}{default}(object, ...)

\method{resample_betas}{fit_clogit}(object, ...)
}
\arguments{
\item{object}{\verb{[a fitted model]} \cr A fitted RSF or (i)SSF model. For possible
models, see \code{\link{supported_models}()}}

\item{...}{Further arguments, none implemented.}
}
\description{
Resamples betas for parametric bootstrap for any supported model type
}
\details{
The \code{default} method is confirmed to work for:
\itemize{
\item \code{glm}
\item \code{clogit}
\item \code{gam} (not tested with actual splines yet)
}
}
