% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{sampling_period}
\alias{sampling_period}
\title{Extract sampling period}
\usage{
sampling_period(x, ...)
}
\arguments{
\item{x}{\verb{[track_xyt]} A \code{track_xyt} object.}

\item{...}{Further arguments, none implemented.}
}
\description{
Extracts sampling period from a \code{track_xyt} object
}
