% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_relocations.R
\name{screen_relocations}
\alias{screen_relocations}
\title{Screen relocations}
\usage{
screen_relocations(x, median_crit, mean_crit, spike_speed, spike_cos, window)
}
\arguments{
\item{x}{A \code{track_xyt}}

\item{median_crit}{\code{[numeric(1)]{>0}} \cr Threshold for the median (in m).}

\item{mean_crit}{\code{[numeric(1)]{>0}} \cr Threshold for the mean (in m).}

\item{spike_speed}{\code{[numeric(1)]{>0}} \cr Threshold spike speed in m/s.}

\item{spike_cos}{\code{[numeric(1)]{>0}} \cr Threshold for the cosine.}

\item{window}{\code{[numeric(1)=10]{>2}} \cr Window width.}
}
\description{
Function to screen relocations for possible errors. The algorithm was first described by Bjorneraas et al. 2010.
}
\examples{
data(deer)
screen_relocations(deer, 100, 100, 0.1, 0.3, 10)
}
