% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fisher.R, R/data-sh.R
\docType{data}
\name{sh}
\alias{sh}
\title{Relocations of 1 red deer}
\format{
A data frame with 1500 rows and 4 variables:
\describe{
\item{x_epsg31467}{the x-coordinate}
\item{y_epsg31467}{the y-coordinate}
\item{day}{the day of the relocation}
\item{time}{the hour of the relocation}
}

A data frame with 1500 rows and 4 variables:
\describe{
\item{x_epsg31467}{the x-coordinate}
\item{y_epsg31467}{the y-coordinate}
\item{day}{the day of the relocation}
\item{time}{the hour of the relocation}
}
}
\source{
Verein für Wildtierforschung Dresden und Göttingen e.V.

Verein für Wildtierforschung Dresden und Göttingen e.V.
}
\usage{
sh

sh
}
\description{
1500 GPS relocations of one red deer from northern Germany.

1500 GPS relocations of one red deer from northern Germany.
}
\keyword{datasets}
