% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_kernel.R
\name{simulate}
\alias{simulate}
\alias{dispersal_kernel}
\alias{simulate_track}
\title{Function to calculate a dispersal kernel, given covariates and coefficients.}
\usage{
dispersal_kernel(
  wx,
  coefficients,
  start,
  spatial.covars,
  temporal.covars = NULL,
  direction = 0,
  max.dist = 100,
  return.raster = FALSE,
  normalize = TRUE
)

simulate_track(
  wx,
  coefficients,
  start,
  spatial.covars,
  direction = 0,
  temporal.covars = NULL,
  max.dist = 100,
  n = 10,
  as.track = TRUE,
  start.time = ymd_hms("2020-01-01 00:00:00"),
  delta.time = hours(2)
)
}
\arguments{
\item{wx}{A fromula with the selection function.}

\item{coefficients}{The coefficient value for each term in \code{wx}.}

\item{start}{The coordinates of the center of the disspersal kernel.}

\item{spatial.covars}{A \code{rasterStack} with the spatial covariates.}

\item{temporal.covars}{A \code{tibble} with non spatial covariates.}

\item{direction}{The direction of the kernel (0 being north) in clockwise direction.}

\item{max.dist}{Maximum dispersal distance.}

\item{return.raster}{Should a raster be returned?}

\item{normalize}{Should the raste be normalized at the end?}

\item{n}{Number of time steps.}

\item{as.track}{Should a track be returned.}

\item{start.time}{The time stamp of the first simulated location.}

\item{delta.time}{Time increment for each subsequent step.}
}
\description{
Function to calculate a dispersal kernel, given covariates and coefficients.
}
