% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_path}
\alias{simulate_path}
\alias{simulate_path.default}
\alias{simulate_path.redistribution_kernel}
\title{Simulate a movement trajectory.}
\usage{
simulate_path(x, ...)

\method{simulate_path}{default}(x, ...)

\method{simulate_path}{redistribution_kernel}(x, n.steps = 100, start = x$args$start, verbose = FALSE, ...)
}
\arguments{
\item{x}{\verb{[redstirubtion_kernel(1)]} \cr An object of class \code{redistribution_kernel}.}

\item{...}{Further arguments, none implemented.}

\item{n.steps}{\verb{[integer(1)]\{100\}} \cr The number of simulation steps.}

\item{start}{\verb{[sim_start]} \cr The starting point in time and space for the simulations (see \code{make_start()}).}

\item{verbose}{\verb{[logical(1)]\{FALSE\}} If \code{TRUE} progress of simulations is displayed.}
}
\description{
Function to simulate a movement trajectory (path) from a redistribution kernel.
}
