% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement2.R
\name{simulate_ud}
\alias{simulate_ud}
\title{Simulate a UD from a dispersal kernel}
\usage{
simulate_ud(movement_kernel, habitat_kernel, start, n = 100000L)
}
\arguments{
\item{n}{Number of time steps}

\item{obj}{A dispersal kernel}

\item{other.vars}{other covariates for each time step.}
}
\description{
Simulate a UD from a dispersal kernel
}
