% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement2.R
\name{simulate_ud_from_dk}
\alias{simulate_ud_from_dk}
\title{Simulate a UD from a dispersal kernel}
\usage{
simulate_ud_from_dk(obj, n = 1000, other.vars = NULL)
}
\arguments{
\item{obj}{A dispersal kernel}

\item{n}{Number of time steps}

\item{other.vars}{other covariates for each time step.}
}
\description{
Simulate a UD from a dispersal kernel
}
