% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_movement2.R
\name{simulate_xy}
\alias{simulate_xy}
\title{Simulate a trajectory}
\usage{
simulate_xy(obj, n = 100, other.vars = NULL)
}
\arguments{
\item{obj}{A dispersal kernel.}

\item{n}{Number of time steps.}

\item{other.vars}{Other covariates (for each time step).}
}
\description{
Simulate a trajectory
}
