% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{sl_distr}
\alias{sl_distr}
\alias{sl_distr.list}
\alias{sl_distr.fit_clogit}
\alias{sl_distr.random_steps}
\title{Step-length distribution}
\usage{
sl_distr(x, ...)

\method{sl_distr}{list}(x, ...)

\method{sl_distr}{fit_clogit}(x, ...)

\method{sl_distr}{random_steps}(x, ...)
}
\arguments{
\item{x}{An object that contains either a fitted conditional logistic regression, random steps or just a fitted distribution.}

\item{...}{Further arguments, none implemented.}
}
\description{
Returns the name of the distribution (e.g., gamma) fitted to the distribution of step lengths.
}
