% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{speed}
\alias{speed}
\title{Speed}
\usage{
speed(x, ...)
}
\arguments{
\item{x}{A \code{track_xyt}.}

\item{append_na}{\code{[logical(1)=TRUE]} \cr Should an \code{NA} be appended at the end.}
}
\value{
\code{[numeric]} \cr The speed in \code{m/s}.
}
\description{
Obtain the speed of a track.
}
