% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{ssf_formula}
\alias{ssf_formula}
\title{Takes a \code{clogit} formula and returns a formula without the \code{strata} and the
left-hand side}
\usage{
ssf_formula(formula)
}
\arguments{
\item{formula}{A formula object}
}
\description{
Takes a \code{clogit} formula and returns a formula without the \code{strata} and the
left-hand side
}
\examples{
f1 <- case_ ~ x1 * x2 + strata(step_id_)
ssf_formula(f1)
}
