% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{ssf_weights}
\alias{ssf_weights}
\title{Given a fitted ssf, and new location the weights for each location is
calculated}
\usage{
ssf_weights(xy, object, compensate.movement = FALSE)
}
\arguments{
\item{xy}{The new locations.}

\item{object}{The the fitted (i)SSF.}

\item{compensate.movement}{Whether or not for the transformation from polar
to Cartesian coordinates is corrected.}
}
\description{
Given a fitted ssf, and new location the weights for each location is
calculated
}
