% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{summarize_sl}
\alias{summarize_sl}
\title{Summarize step lengths}
\usage{
summarize_sl(x, ...)
}
\arguments{
\item{x}{\verb{[track_xy, track_xyt]} A \verb{track_xy*} object.}

\item{...}{Further arguments, none implemented.}
}
\description{
Summarizes step lengths for a \verb{track_xy*} object
}
