% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{summarize_speed}
\alias{summarize_speed}
\title{Summarize speed}
\usage{
summarize_speed(x, ...)
}
\arguments{
\item{x}{\verb{[track_xyt]} A \code{track_xyt} object.}
}
\description{
Summarizes speeds for a \code{track_xyt} object
}
