% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_support.R
\name{supported_models}
\alias{supported_models}
\title{Currently supported models}
\usage{
supported_models()
}
\value{
Returns list of possible RSF and (i)SSF models invisibly.
}
\description{
Prints currently supported models
}
\examples{

supported_models()


}
\author{
Brian J. Smith
}
