% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{ta_distr}
\alias{ta_distr}
\alias{ta_distr.list}
\alias{ta_distr.fit_clogit}
\alias{ta_distr.random_steps}
\title{Turn angle distribution}
\usage{
ta_distr(x, ...)

\method{ta_distr}{list}(x, ...)

\method{ta_distr}{fit_clogit}(x, ...)

\method{ta_distr}{random_steps}(x, ...)
}
\arguments{
\item{x}{\code{[fit_clogit(1),random_steps(1)]} \cr An object that contains either a fitted conditional logistic regression, random steps or just a fitted distribution.}

\item{...}{Further arguments, none implemented.}
}
\description{
Returns the name of the distribution (e.g., von Mises) fitted to the distribution of turn angles.
}
