% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{ta_params}
\alias{ta_params}
\alias{ta_params.fit_clogit}
\alias{ta_kappa}
\title{Turn angle parameters}
\usage{
ta_params(x, ...)

\method{ta_params}{fit_clogit}(x, ...)

ta_kappa(x, ...)
}
\arguments{
\item{x}{\code{[fit_clogit(1),random_steps(1)]} \cr An object that contains either a fitted conditional logistic regression, random steps or just a fitted distribution.}

\item{...}{Further arguments, none implemented.}
}
\description{
Returns the parameter of the distribution (e.g., von Mises) fitted to the distribution of turn angles.
}
