% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Create a Track}
\usage{
track(x, y, t, ..., crs = NULL)
}
\arguments{
\item{x}{A \emph{required} numeric vector, containing the x-coordinates (=
latitude).}

\item{y}{A \emph{required} numeric vector, containing the y-coordinates (=
longitude).}

\item{t}{An \emph{optional} vector of  \code{POSIXct} time stamps.}

\item{...}{Additional columns, that should be given in the form of \code{key = val} (e.g., for ids this may look like this \code{id = c(1, 1, 1, 2, 2, 2} for
three points for ids 1 and 2 each).}

\item{crs}{An optional coordinate reference system of the points.}
}
\value{
If \code{t} was provided an object of class \code{track_xyt} is returned
otherwise a \code{track_xy}.
}
\description{
Consructor to crate a track, the basic building block of the \code{amt} pacakge. A
\code{track} is usually created from a set of \code{x} and \code{y} coordinates, possibly
time stamps, and any number of optional columns, such as id, sex, age, etc.
}
