% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_align.R
\name{track_align}
\alias{track_align}
\alias{track_align.track_xyt}
\title{Selects relocations that fit a new time series}
\usage{
track_align(x, ...)

\method{track_align}{track_xyt}(x, new.times, tolerance, ...)
}
\arguments{
\item{x}{A track.}

\item{...}{Further arguments, none implemented.}

\item{new.times}{The new time trajectory.}

\item{tolerance}{The tolerance between observed time stamps and new time stamps in seconds. This should be either a vector of length 1 or length \code{new.times}.}
}
\value{
A \code{track_xyt}.
}
\description{
Functions to only selects relocations that can be aligned with a new time series (within some tolerance).
}
