% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{track_methods}
\alias{track_methods}
\alias{velocity}
\alias{velocity.track_xyt}
\alias{nsd}
\alias{nsd.track_xy}
\alias{diff_x}
\alias{diff_x.track_xy}
\alias{diff_y}
\alias{diff_y.track_xy}
\alias{direction_abs}
\alias{direction_abs.track_xy}
\alias{direction_rel}
\alias{direction_rel}
\alias{step_lengths}
\alias{step_lengths.track_xy}
\alias{step_lenths_sq}
\alias{step_lengths_sq}
\alias{centroid}
\alias{centroid.track_xy}
\title{Track Methods}
\usage{
velocity(x, ...)

\method{velocity}{track_xyt}(x, ...)

nsd(x, ...)

\method{nsd}{track_xy}(x, ...)

diff_x(x, ...)

\method{diff_x}{track_xy}(x, ...)

diff_y(x, ...)

\method{diff_y}{track_xy}(x, ...)

direction_abs(x)

\method{direction_abs}{track_xy}(x)

direction_rel(x)

direction_rel(x)

step_lengths(x, ...)

\method{step_lengths}{track_xy}(x, ...)

step_lenths_sq(x, ...)

step_lengths_sq(x)

centroid(x, ...)

\method{centroid}{track_xy}(x, ...)
}
\arguments{
\item{x}{A track_xy{t}.}

\item{...}{Further arguments, none implemented.}
}
\description{
Methods to work with a track.
Function to calculate the absolute direction of a movement track. 0 is north.
}
