% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_resample.R
\name{track_resample}
\alias{track_resample}
\alias{track_resample.track_xyt}
\title{Resample track}
\usage{
track_resample(x, ...)

\method{track_resample}{track_xyt}(x, rate = hours(2), tolerance = minutes(15), start = 1, ...)
}
\arguments{
\item{x}{A \code{track_xyt}.}

\item{...}{Further arguments, none implemented.}

\item{rate}{A lubridate \code{Period}, that indicates the sampling rate.}

\item{tolerance}{A lubridate \code{Period}, that indicates the tolerance of deviations of the sampling rate.}

\item{start}{A integer scalar, that gives the relocation at which the sampling rate starts.}
}
\description{
Function to resample a track at a predefined sampling rate within some tolerance.
}
