% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_transform_crs.R
\name{transform_coords}
\alias{transform_coords}
\alias{transform_coords.track_xy}
\title{Transform CRS}
\usage{
transform_coords(x, ...)

\method{transform_coords}{track_xy}(x, crs_to, crs_from)
}
\arguments{
\item{x}{A track.}

\item{crs_to}{coordinate reference system the data should be transformed to, see \code{sp::CRS}.}

\item{crs_from}{coordinate reference system the data are currently in, see \code{sp::CRS}.}
}
\description{
Transforms the CRS for a track.
}
\seealso{
sp::spTransform
}
