% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_kernel.R
\name{Transformations}
\alias{Transformations}
\alias{kappa_to_cos_ta}
\alias{scale_to_sl}
\alias{shape_to_log_sl}
\alias{sl_to_scale}
\alias{log_sl_to_shape}
\alias{cos_ta_to_kappa}
\title{Functions to transform parameters for the movement model (from the
distributions that describe the turn angles and step lengths) to coefficients
of the selection function.}
\usage{
kappa_to_cos_ta(param)

scale_to_sl(param)

shape_to_log_sl(param)

sl_to_scale(coef)

log_sl_to_shape(coef)

cos_ta_to_kappa(coef)
}
\arguments{
\item{param}{A parameter of a distribution.}

\item{coef}{Name of the coefficient.}
}
\value{
A named vector of length one with the coefficient for the selection function.
}
\description{
Functions to transform parameters for the movement model (from the
distributions that describe the turn angles and step lengths) to coefficients
of the selection function.
}
