% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-uhc.R
\docType{data}
\name{uhc_hab}
\alias{uhc_hab}
\title{Simulated habitat rasters for demonstrating UHC plots}
\format{
A \code{RasterStack} with 1600 cells and 7 variables:
\describe{
\item{forage}{Forage biomass in g/m^2^ (resource)}
\item{temp}{mean annual temperature in °C (condition)}
\item{pred}{predator density in predators/100 km^2^ (risk)}
\item{cover}{landcover (forest > grassland > wetland)}
\item{dist_to_water}{distance to the wetland landcover (no effect)}
\item{dist_to_cent}{distance to the centroid of the raster (no effect)}
\item{rand}{random integers (no effect)}
}
}
\usage{
uhc_hab
}
\description{
Simulated habitat rasters for demonstrating UHC plots
}
\keyword{datasets}
