% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_parameters.R
\name{adjust_param}
\alias{adjust_param}
\alias{adjust_shape}
\alias{adjust_scale}
\alias{adjust_kappa}
\title{Adjust parameters}
\usage{
adjust_shape(tentative, modifier = 0)

adjust_scale(tentative, modifier = 0)

adjust_kappa(tentative, modifier = 0)
}
\arguments{
\item{tentative}{\code{[numeric]} \cr The tentative parameter estimate.}

\item{modifier}{\code{[numeric=0]} \cr The modifier to adjust the tentative
estimate.}
}
\description{
Functions aiding parameter adjustment after fitting an integrated step
selection function (iSSF).
}
\details{
Currently the shape and scale parameter of a Gamma distribution and the
concentration parameter (=kappa) of a von-Mises distribution can be adjusted.
When fitting a integrated step selection model, the tentative shape parameter
of the Gamma distribution of the step length can be adjusted with the
coefficients of the log of step lengths, the scale of the Gamma distribution
can be adjusted with the step lengths, and the kappa parameter of the von
Mises distribution can be adjusted with the coefficient for the cosine of the
turn angle.
}
\references{
\insertRef{avgar2016}{amt}
}
