% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amt.R
\docType{package}
\name{amt-package}
\alias{amt}
\alias{amt-package}
\title{amt: Animal Movement Tools}
\description{
Manage and analyze animal movement data. The functionality of 'amt' includes methods to calculate track statistics (e.g. step lengths, speed, or turning angles), prepare data for fitting habitat selection analyses (resource and step-selection functions), and simulation of space-use from fitted step-selection functions.
}
\author{
\strong{Maintainer}: Johannes Signer \email{jsigner@gwdg.de}

Other contributors:
\itemize{
  \item Bjoern Reineking [contributor]
  \item Scott LaPoint [data contributor]
}

}
