% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amt.R
\docType{package}
\name{amt-package}
\alias{amt}
\alias{amt-package}
\title{amt: Animal Movement Tools}
\description{
Manage and analyze animal movement data. The functionality of 'amt' includes methods to calculate home ranges <doi:10.1101/2020.08.19.256859v2>, track statistics (e.g. step lengths, speed, or turning angles), prepare data for fitting habitat selection analyses <doi:10.1101/2020.11.12.379834v4>, and simulation of space-use from fitted step-selection functions <doi:10.1002/ecs2.1771>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmsigner/amt}
}

}
\author{
\strong{Maintainer}: Johannes Signer \email{jsigner@gwdg.de}

Other contributors:
\itemize{
  \item Brian Smith [contributor]
  \item Bjoern Reineking [contributor]
  \item Ulrike Schlaegel [contributor]
  \item John Fieberg [contributor]
  \item Josh O'Brien [contributor]
  \item Bernardo Niebuhr [contributor]
  \item Scott LaPoint [data contributor]
}

}
