% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amt.R
\docType{package}
\name{amt-package}
\alias{amt}
\alias{amt-package}
\title{amt: Animal Movement Tools}
\description{
Manage and analyze animal movement data. The functionality of 'amt' includes methods to calculate track statistics (e.g. step lengths, speed, or turning angles), prepare data for fitting habitat selection analyses (resource selection functions and step-selection functions <doi:10.1890/04-0953> and integrated step-selection functions <doi:10.1111/2041-210X.12528>), and simulation of space-use from fitted step-selection functions <doi:10.1002/ecs2.1771>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmsigner/amt}
}

}
\author{
\strong{Maintainer}: Johannes Signer \email{jsigner@gwdg.de}

Other contributors:
\itemize{
  \item Brian Smith [contributor]
  \item Bjoern Reineking [contributor]
  \item Ulrike Schlaegel [contributor]
  \item Scott LaPoint [data contributor]
}

}
