% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fisher.R
\docType{data}
\name{amt_fisher_covar}
\alias{amt_fisher_covar}
\title{Environmental data for fishers}
\format{
A list with three where each entry is a \code{RasterLayer}.
}
\source{
\verb{https://lpdaac.usgs.gov/dataset_discovery/aster/aster_products_table}

\verb{http://dup.esrin.esa.it/page_globcover.php}

\verb{http://sedac.ciesin.columbia.edu/data/collection/gpw-v3/sets/browse}
}
\usage{
amt_fisher_covar
}
\description{
A list with three entries that correspond to the following three layer: land
use, elevation and population density.
}
\keyword{datasets}
