% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rss.R
\name{append_x1}
\alias{append_x1}
\title{Append "_x1"}
\usage{
append_x1(string)
}
\arguments{
\item{string}{\verb{[character]} Variable name to possibly append to}
}
\description{
Helper function to append "_x1" to variable names
}
\details{
The function first checks if "_x1" is already appended and adds it if
it is not. This is meant for internal use in \verb{\link{plot.log_rss}()}.
}
