% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_track.R
\name{as_track}
\alias{as_track}
\alias{as_track.SpatialPoints}
\title{Coerce to track}
\usage{
as_track(x, ...)

\method{as_track}{SpatialPoints}(x, ...)
}
\arguments{
\item{x}{\code{[SpatialPoints]} \cr Object to be converted to a track.}

\item{...}{Further arguments, none implemented.}
}
\description{
Coerce other classes (currently implemented: \code{SpatialPoints}) to a \code{track_xy}.
}
\examples{
xy <- sp::SpatialPoints(cbind(c(1, 3, 2, 1), c(3, 2, 2, 1)))
as_track(xy)
}
