% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_track.R
\name{as_track}
\alias{as_track}
\alias{as_track.sfc_POINT}
\alias{as_track.steps_xyt}
\alias{as_track.data.frame}
\title{Coerce to track}
\usage{
as_track(x, ...)

\method{as_track}{sfc_POINT}(x, ...)

\method{as_track}{steps_xyt}(x, ...)

\method{as_track}{data.frame}(x, ...)
}
\arguments{
\item{x}{Object to be converted to a track.}

\item{...}{Further arguments, none implemented.}
}
\value{
An object of class \code{track_xy(t)}
}
\description{
Coerce other classes to a \code{track_xy}.
}
