% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{bbox}
\alias{bbox}
\alias{bbox.track_xy}
\title{Get bounding box of a track}
\usage{
bbox(x, ...)

\method{bbox}{track_xy}(x, spatial = TRUE, buffer = NULL, ...)
}
\arguments{
\item{x}{A track.}

\item{...}{Further arguments, none implemented.}

\item{spatial}{Logical, whether or not to return a \code{sp}-object.}

\item{buffer}{Numeric, an optional buffer.}
}
\description{
Get bounding box of a track
}
\examples{
data(deer)
bbox(deer)
bbox(deer, buffer = 100)
}
