% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rss.R
\name{boot1.fit_logit}
\alias{boot1.fit_logit}
\alias{boot1.fit_clogit}
\title{Single bootstrap iteration}
\usage{
boot1.fit_logit(object, x1, x2)

boot1.fit_clogit(object, x1, x2)
}
\description{
Runs a single iteration of the empirical bootstrap
}
\details{
This function is meant for internal use by \code{bootstrap_logrss()} and
is not meant to be called by the user.
}
\keyword{internal}
