% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{centroid}
\alias{centroid}
\title{Centroid of a track}
\usage{
centroid(x, ...)
}
\arguments{
\item{x}{A track.}

\item{...}{Further arguments, none implemented.}
}
\description{
Calcualte the centroid of a track.
}
\examples{
data(deer)
centroid(deer)
}
