% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{centroid}
\alias{centroid}
\alias{centroid.track_xy}
\title{Calculate the centroid of a track.}
\usage{
centroid(x, ...)

\method{centroid}{track_xy}(x, spatial = FALSE, ...)
}
\arguments{
\item{x}{\code{[track_xy, track_xyt]} \cr A track created with \code{make_track}.}

\item{...}{Further arguments, none implemented.}

\item{spatial}{\verb{[logical(1)=FALSE]} \cr Whether or not to return a \code{SpatialPoints}-object.}
}
\description{
Calculate the centroid of a track.
}
\examples{
data(deer)
centroid(deer)
}
