% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_rss.R, R/log_rss_old.R
\name{check_factors}
\alias{check_factors}
\title{Check factor levels}
\usage{
check_factors(model, x1, x2)

check_factors(model, x1, x2)
}
\arguments{
\item{model}{\verb{[glm, clogit]} \cr The model object from a fitted RSF/(i)SSF.
\emph{I.e.}, it will be \code{object$model} when called within \code{log_rss()}.}

\item{x1}{\verb{[data.frame]} \cr A \code{data.frame} representing the habitat values
at location x_1. Must contain all fitted covariates as expected by
\code{predict()}.}

\item{x2}{\verb{[data.frame]} \cr A 1-row \code{data.frame} representing the single
hypothetical location of x_2. Must contain all fitted covariates as expected
by \code{predict()}.}
}
\description{
Check factor levels before log-RSS calculation

Check factor levels before log-RSS calculation
}
\details{
This function is meant for internal use by \code{log_rss()} and is
not meant to be called by the user.

This function is meant for internal use by \code{log_rss()} and is
not meant to be called by the user.
}
\keyword{internal}
