% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{check_time_unit}
\alias{check_time_unit}
\title{Calculate Speed}
\usage{
check_time_unit(tu)
}
\description{
Calculates speed
}
\details{
Calculate Change in NSD

Calculates change in NSD

Check \code{time_unit} Parameter

Internal function to check \code{time_unit} parameter in various cleaning functions.
}
