% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{coercion}
\alias{coercion}
\alias{as_sf}
\alias{as_sf.steps_xy}
\alias{as_sp}
\alias{as_ltraj}
\alias{as_ltraj.track_xy}
\alias{as_ltraj.track_xyt}
\alias{as_telemetry}
\alias{as_telemetry.track_xyt}
\alias{as_moveHMM}
\alias{as_moveHMM.track_xy}
\title{Coerce a track to other formats.}
\usage{
as_sf(x, ...)

\method{as_sf}{steps_xy}(x, end = TRUE, ...)

as_sp(x, ...)

as_ltraj(x, ...)

\method{as_ltraj}{track_xy}(x, id = "animal_1", ...)

\method{as_ltraj}{track_xyt}(x, ...)

as_telemetry(x, ...)

\method{as_telemetry}{track_xyt}(x, ...)

as_moveHMM(x, ...)

\method{as_moveHMM}{track_xy}(x, ...)
}
\arguments{
\item{x}{\code{[track_xy, track_xyt]} \cr A track created with \code{make_track}.}

\item{...}{Further arguments, none implemented.}

\item{end}{\verb{[logical(1)=TRUE]} \cr For steps, should the end or start points be used?}

\item{id}{\verb{[numeric,character,factor]} \cr Animal id(s).}
}
\value{
An object of the class to which coercion is performed to.
}
\description{
Several other packages provides methods to analyze movement data, and \code{amt} provides coercion methods to some packages.
}
