% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{coercion}
\alias{coercion}
\alias{as_sp}
\alias{as_sp.track_xy}
\alias{as_move}
\alias{as_move.track_xy}
\alias{as_move.track_xyt}
\title{Track coercion functions}
\usage{
as_sp(x, ...)

\method{as_sp}{track_xy}(x, ...)

as_move(x, ...)

\method{as_move}{track_xy}(x, ...)

\method{as_move}{track_xyt}(x, ...)
}
\arguments{
\item{x}{A \code{track_xy}.}

\item{...}{Further arguments, none implemented.}
}
\description{
amt provides several coercion functions that operate on \code{tracks} and \code{steps}.
}
\details{
The following coercion functions are currently provided:
\itemize{
\item \code{as_sp} turns a track into \code{SpatialPoints}.
}
}
