% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R
\name{coercion}
\alias{coercion}
\alias{as_sp}
\alias{as_sp.steps_xy}
\alias{as_move}
\alias{as_move.track_xyt}
\alias{as_ltraj}
\alias{as_ltraj.track_xy}
\alias{as_ltraj.track_xyt}
\alias{as_bcpa}
\alias{as_bcpa.track_xyt}
\alias{as_telemetry}
\alias{as_telemetry.track_xyt}
\alias{as_moveHMM}
\alias{as_moveHMM.track_xy}
\title{Coerce a track to other formats.}
\usage{
as_sp(x, ...)

\method{as_sp}{steps_xy}(x, end = TRUE, ...)

as_move(x, ...)

\method{as_move}{track_xyt}(x, id = "id", ...)

as_ltraj(x, ...)

\method{as_ltraj}{track_xy}(x, id = "animal_1", ...)

\method{as_ltraj}{track_xyt}(x, ...)

as_bcpa(x, ...)

\method{as_bcpa}{track_xyt}(x, ...)

as_telemetry(x, ...)

\method{as_telemetry}{track_xyt}(x, ...)

as_moveHMM(x, ...)

\method{as_moveHMM}{track_xy}(x, ...)
}
\arguments{
\item{x}{\code{[track_xy, track_xyt]} \cr A track created with \code{make_track}.}

\item{...}{Further arguments, none implemented.}

\item{end}{\verb{[logical(1)=TRUE]} \cr For steps, should the end or start points be used?}

\item{id}{\verb{[numeric,character,factor]} \cr Animal id(s).}
}
\description{
Several other packages provides methods to analyze movement data, and \code{amt} provides coercion methods to some packages
}
\examples{
data(deer)
as_move(deer)
as_move(deer, id = "foo")
data(deer)
as_ltraj(deer)
as_ltraj(deer, id = "animal_3")
data(deer)
d <- as_bcpa(deer)
data(deer)
as_telemetry(deer)
# Fit HMM with two states
data(deer)
dm <- as_moveHMM(deer)
}
