% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_ud.R
\name{cum_ud}
\alias{cum_ud}
\alias{cumulative_ud}
\alias{cumulative_ud.RasterLayer}
\alias{cumulative_ud.kde}
\title{Calculate a cumulative UD}
\usage{
cumulative_ud(x, ...)

\method{cumulative_ud}{RasterLayer}(x, ...)

\method{cumulative_ud}{kde}(x, ...)
}
\arguments{
\item{x}{\verb{[RasterLayer]} \cr Containing the Utilization Distribution (UD).}

\item{...}{Further arguments, none implemented.}
}
\value{
\verb{[RasterLayer]} \cr The cumulative UD.
}
\description{
Calculate the cumulative utilization distribution (UD).
}
\note{
This function is typically used to obtain isopleths.
}
