% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_cud.R
\name{cum_ud}
\alias{cum_ud}
\alias{hr_cud}
\alias{hr_cud.RasterLayer}
\title{Calculate a cumulative UD}
\usage{
hr_cud(x, ...)

\method{hr_cud}{RasterLayer}(x, ...)
}
\arguments{
\item{x}{\verb{[RasterLayer]} \cr Containing the Utilization Distribution (UD).}

\item{...}{Further arguments, none implemented.}
}
\value{
\verb{[RasterLayer]} \cr The cumulative UD.
}
\description{
Calculate the cumulative utilization distribution (UD).
}
\note{
This function is typically used to obtain isopleths.
}
