% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sh.R
\docType{data}
\name{deer}
\alias{deer}
\title{Relocations of 1 red deer}
\format{
A track_xyt \describe{ \item{x_}{the x-coordinate} \item{y_}{the
y-coordinate} \item{t_}{the timestamp} \item{burst_}{the burst a particular
points belongs to.} }
}
\source{
Verein für Wildtierforschung Dresden und Göttingen e.V.
}
\usage{
deer
}
\description{
826 GPS relocations of one red deer from northern Germany. The data is
already resampled to a regular time interval of 6 hours and the coordinate
reference system is transformed to \code{epsg:3035}.
}
\keyword{datasets}
