% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{direction}
\alias{direction}
\alias{direction_abs}
\alias{direction_abs.track_xy}
\alias{direction_rel}
\alias{direction_rel}
\title{Directions}
\usage{
direction_abs(x, ...)

\method{direction_abs}{track_xy}(x, degrees = TRUE, full_circle = FALSE,
  zero_dir = "E", ...)

direction_rel(x, degrees = TRUE)

direction_rel(x, degrees = TRUE)
}
\arguments{
\item{x}{A track.}

\item{degrees}{Logcial scalar, if \code{TRUE} angles are returned in degrees, otherwise in radians.}

\item{full_circle}{Logcial scalar, if \code{TRUE} angles are returned between 0 and 360 degrees or 0 and 2 pi (depnding on the value of \code{degrees}), otherwise angles are between -180 and 180 or -pi and pi.}

\item{zero}{Character scalar, must be either 'N', 'E', 'S' or 'W' and indicates the zero direction.}
}
\description{
Funcitons to calculate the relative and absolute direction of an animal
}
\details{
Anti clockwise angles of steps to an absolute direction (North, East, South or West).
}
\examples{
# Absolute directions

# east
x <- c(0, 1)
y <- c(0, 0)
trk <- track(x = x, y = y)
direction_abs(trk)

# south
x <- c(0, 0)
y <- c(0, -1)
trk <- track(x = x, y = y)
direction_abs(trk)

# west
x <- c(0, -1)
y <- c(0, 0)
trk <- track(x = x, y = y)
direction_abs(trk)

# north
x <- c(0, 0)
y <- c(0, 1)
trk <- track(x = x, y = y)
direction_abs(trk)

# ne
x <- c(0, 1)
y <- c(0, 1)
trk <- track(x = x, y = y)
direction_abs(trk)

# se
x <- c(0, 1)
y <- c(0, -1)
trk <- track(x = x, y = y)
direction_abs(trk)

# sw
x <- c(0, -1)
y <- c(0, -1)
trk <- track(x = x, y = y)
direction_abs(trk)

# nw
x <- c(0, -1)
y <- c(0, 1)
trk <- track(x = x, y = y)
direction_abs(trk)

# real data
data(sh)
trk <- track(x = sh$x, y = sh$y)
\dontrun{
hist(direction_abs(trk))
}
}
