% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_clogit.R
\name{fit_clogit}
\alias{fit_clogit}
\title{Fit a conditional logistic regression}
\usage{
fit_clogit(data, formula, more = NULL, ...)
}
\arguments{
\item{data}{The data.}

\item{formula}{The formula}

\item{more}{Optional list returned in the output.}

\item{...}{Addtional arguments, passed to \code{survival::clogit}.}
}
\description{
Tihs function is a wrapper around \code{stats::glm}, making it usable in a piped workflow.
}
