% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_clogit.R
\name{fit_clogit}
\alias{fit_clogit}
\title{Fit a conditional logistic regression}
\usage{
fit_clogit(data, formula, more = NULL, summary_only = FALSE, ...)
}
\arguments{
\item{data}{\code{[data.frame]} \cr The data used to fit a model.}

\item{formula}{\code{[formula]} \cr The model formula.}

\item{more}{\code{[list]} \cr Optional list that is passed on the output.}

\item{summary_only}{\code{[logical(1)=FALSE]} \cr If \code{TRUE} only a \code{broom::tidy} summary of the model is returned.}

\item{...}{Additional arguments, passed to \code{survival::clogit}.}
}
\description{
This function is a wrapper around \code{stats::glm}, making it usable in a piped workflow.
}
