% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_clogit.R
\name{fit_clogit}
\alias{fit_clogit}
\title{Fit a conditional logistic regression}
\usage{
fit_clogit(data, formula, more = NULL, summary_only = FALSE, ...)
}
\arguments{
\item{data}{The data.}

\item{formula}{The formula}

\item{more}{Optional list returned in the output.}

\item{summary_only}{\code{[logical(1)=FALSE]} \cr If \code{TRUE} only a \code{broom::tidy} summary of the model is returned.}

\item{...}{Addtional arguments, passed to \code{survival::clogit}.}
}
\description{
This function is a wrapper around \code{stats::glm}, making it usable in a piped workflow.
}
\details{
When including step length and log(step length) in the model and later parameter adjustment is desired, step length must be called \code{sl_} and log(step length) either \code{log_sl_} when precalculated, or \code{log(sl_)} when it is calculated within the formula.
}
