% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_logit.R
\name{fit_logit}
\alias{fit_logit}
\alias{fit_rsf}
\title{Fit logistic regression}
\usage{
fit_logit(data, formula, ...)

fit_rsf(data, formula, ...)
}
\arguments{
\item{data}{\verb{[data.frame]} \cr The data used to fit a model.}

\item{formula}{\verb{[formula]} \cr The model formula.}

\item{...}{Further arguments passed to \code{stats::glm}.}
}
\description{
This function is a wrapper around \code{stats::glm} for piped workflows.
}
