% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{fit_ta_dist}
\alias{fit_ta_dist}
\alias{fit_ta_dist_base}
\title{Fit a statistical distribution to the turn angles of a track}
\usage{
fit_ta_dist(.tbl, x, ...)

fit_ta_dist_base(x, na.rm = TRUE, distr = "vonmises", ...)
}
\arguments{
\item{.tbl}{\code{[track_xy,track_xyt]} \cr A track.}

\item{x}{\code{[expression]} \cr The name of the column containing turn angles, usually \code{ta_}.}

\item{...}{Further arguments, none implemented.}

\item{na.rm}{\code{[logical(1)]} \cr Should \code{NA} be removed?}

\item{distr}{\code{[character(1)]{'vonmises', 'unif'}} \cr Name of the distribution, currently only \code{vonmises} and \code{uniform} distribution is supported.}
}
\description{
Fit a statistical distribution to the turn angles of a track
}
