% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{get_crs}
\alias{get_crs}
\title{Coordinate References System}
\usage{
get_crs(x, ...)
}
\arguments{
\item{x}{A track.}

\item{...}{Further arguments, none implemented.}
}
\description{
Get the Coordinate Reference System (CRS) of a track.
}
