% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{get_kappa}
\alias{get_kappa}
\title{Get \code{kappa}}
\usage{
get_kappa(x, ...)
}
\arguments{
\item{x}{\code{[list]} \cr Fitted turn angle distribution.}

\item{...}{Further arguments, none implemented.}
}
\description{
Convenience function to extract kappa and its SE.
}
