% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_area.R, R/hr_isopleths.R
\name{hr}
\alias{hr}
\alias{hr_area}
\alias{hr_area.mcp}
\alias{hr_area.locoh}
\alias{hr_area.RasterLayer}
\alias{hr_isopleths}
\alias{hr_isopleths.RasterLayer}
\alias{hr_isopleths.mcp}
\title{Working with home ranges}
\usage{
hr_area(x, ...)

\method{hr_area}{mcp}(x, ...)

\method{hr_area}{locoh}(x, ...)

\method{hr_area}{RasterLayer}(x, levels = 0.95, ...)

hr_isopleths(x, ...)

\method{hr_isopleths}{RasterLayer}(x, levels = 0.95, ...)

\method{hr_isopleths}{mcp}(x, ...)
}
\arguments{
\item{x}{Fitted home range.}

\item{...}{Further arguments, none implemented.}

\item{levels}{Numeric vector, the home range level (\code{0 < level < 1} is assumed).}
}
\description{
Function to calculate the area and isopleths of a home range.
}
