% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_overlap.R
\name{hr_overlaps}
\alias{hr_overlaps}
\alias{hr_overlap}
\alias{hr_ba}
\title{Different Methods to calculate home-range overlaps}
\usage{
hr_overlap(x, ...)

hr_ba(x, ...)
}
\arguments{
\item{x}{\code{hr}}

\item{...}{Further arguments, none implemented.}

\item{y}{\code{hr}}
}
\value{
\code{data.frame} with the isopleth level and area in units of the coordinate reference system.
}
\description{
Different Methods to calculate home-range overlaps
}
