% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_overlap.R
\name{hr_overlaps}
\alias{hr_overlaps}
\alias{hr_overlap}
\alias{hr_overlap.hr}
\alias{hr_overlap.list}
\title{Different Methods to calculate home-range overlaps}
\usage{
hr_overlap(x, ...)

\method{hr_overlap}{hr}(x, y, ...)

\method{hr_overlap}{list}(x, consecutive.only = TRUE, labels = NULL, ...)
}
\arguments{
\item{x, y}{\code{hr} A home-range estimate}

\item{...}{Further arguments, none implemented.}

\item{consecutive.only}{\verb{[logical=TRUE]} \cr Should only consecutive overlaps be calculated or all combinations?}

\item{labels}{\verb{[character=NULL]} \cr Labels for different instances. If \code{NULL} (the default) numbers will be used.}
}
\value{
\code{data.frame} with the isopleth level and area in units of the coordinate reference system.
}
\description{
Different Methods to calculate home-range overlaps
}
