% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_ud.R
\name{hr_ud}
\alias{hr_ud}
\title{Obtain the utilization distribution of a probabilistic home range}
\usage{
hr_ud(x, ...)
}
\arguments{
\item{x}{\verb{[hr_prob]} The home-range estimate}

\item{...}{Further arguments, none implemented.}
}
\description{
Obtain the utilization distribution of a probabilistic home range
}
