% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_mcp.R
\name{mcp}
\alias{mcp}
\alias{hr_mcp}
\alias{hr_mcp.track_xy}
\title{Minimum Convex Polygon}
\usage{
hr_mcp(x, levels = 0.95, ...)

\method{hr_mcp}{track_xy}(x, levels = 0.95, ...)
}
\arguments{
\item{x}{A track.}

\item{levels, }{the home range levels.}

\item{...}{Further arguments, none implemented.}
}
\description{
Function to calcualte a Minimum Convex Polygon (MCP) home range.
}
\examples{
data(sh)
trk <- track(sh$x_, sh$y_)
mcps <- mcp(trk, levels = c(0.5, 0.95, 1))
\dontrun{
plot(trk, asp = 1)
sp::plot(hr_mcps$mcp, add = TRUE)
}


}
