% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_mcp.R
\name{mcp}
\alias{mcp}
\alias{mcp.track_xy}
\title{Minimum Convex Polygon}
\usage{
mcp(x, levels = 0.95, ...)

\method{mcp}{track_xy}(x, levels = 0.95, ...)
}
\arguments{
\item{x}{A track.}

\item{levels, }{the home range levels.}

\item{...}{Further arguments, none implemented.}
}
\description{
Function to calcualte a Minimum Convex Polygon (MCP) home range.
}
