% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{prep_uhc}
\alias{prep_uhc}
\alias{prep_uhc.glm}
\title{Prepare Data for UHC Plots for a Fitted Model}
\usage{
prep_uhc(object, test_dat, n_samp = 1000, verbose = TRUE)

\method{prep_uhc}{glm}(object, test_dat, n_samp = 1000, verbose = TRUE)
}
\arguments{
\item{object}{\verb{[glm, fit_logit, fit_clogit]} \cr A fitted RSF or (i)SSF model.
Should be fit to \emph{training} dataset separate from the testing data.}

\item{test_dat}{\verb{[data.frame]} \cr A \code{data.frame} with \emph{testing} data from
which to sample test points. Should be separate from the data used to train
the model passed to \code{object}.}

\item{n_samp}{\verb{[numeric = 1000]} A \code{vector} of length 1 giving the number of
samples to use to characterize the used habitat distribution under the model.}

\item{verbose}{\verb{[logical]} Should messages be displayed (\code{TRUE}) or not
(\code{FALSE})?}
}
\value{
Returns a \code{list} of class \code{uhc_data} with elements:
\itemize{
\item \code{orig}: List of \code{data.frame}s, one per variable (see \code{vars}). Each
\code{data.frame} contains the density plot data (\code{x} and \code{y}) for the original
used (\code{dist == "U"}) and available (\code{dist == "A"}) data.
\item \code{samp}: List of \code{data.frame}s, one per variable (see \code{vars}). Each
\code{data.frame} contains the density plot data (\code{x} and \code{y}) for each iteration
of bootstrap resampling (\code{iter}).
\item \code{vars}: Character vector with names of the habitat variables for which to
create UHC plots.
\item \code{type}: Named character vector with the type for each of \code{vars} (either
\code{"numeric"} or \code{"factor"}).
\item \code{resp}: Character vector of length 1 with the name of the response
variable.
}
}
\description{
Creates data used to make used-habitat calibration plots
}
\details{
This function performs the heavy lifting of creating UHC plots.
It creates the data used later by the \code{plot()} method, which actually
draws the UHC plots. This function (1) creates density plots of the used
and available locations from the \emph{test} data, and (2) resamples the (a)
fitted coefficients and (b) test data (weighted by the exponential habitat
selection function) to create the distribution of used habitat under the
model.

Note that \code{test_dat} should contain at least all of the variables that
appear in the model \code{object}. Any further habitat variables in \code{test_dat}
will also have UHC plots generated, treating these variables as possible
candidate variables that are simply not included in this particular model.
}
\examples{

\donttest{

# Load data
data(uhc_hsf_locs)
data(uhc_hab)

#

}
}
\references{
Fieberg, J.R., Forester, J.D., Street, G.M., Johnson, D.H., ArchMiller, A.A.,
and Matthiopoulos, J. 2018. Used-habitat calibration plots: A new procedure
for validating species distribution, resource selection, and step-selection
models. \emph{Ecography} 41:737–752.
}
\seealso{
See Fieberg \emph{et al.} 2018 for details about UHC plots.

Default plotting method available: \code{\link{plot.uhc_data}()}
Coercion to \code{data.frame}: \code{\link{as.data.frame.uhc_data}()}
}
