% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{prep_uhc}
\alias{prep_uhc}
\alias{prep_uhc.glm}
\title{Create UHC Plots for a Fitted Model}
\usage{
prep_uhc(object, test_dat)

\method{prep_uhc}{glm}()
}
\arguments{
\item{object}{\verb{[fit_logit, fit_clogit]} \cr A fitted RSF or (i)SSF model.}

\item{test_dat}{\verb{[data.frame]} \cr A \code{data.frame} with \emph{testing} data from
which to sample test points.}
}
\value{
Returns a \code{list} of class \code{uhc_sim} with elements:
\itemize{
\item \code{el1}: TBD
}
}
\description{
Creates used-habitat calibration plots
}
\details{
TBD
}
\examples{

#' \donttest{

}
}
\references{
Fieberg, J.R., Forester, J.D., Street, G.M., Johnson, D.H., ArchMiller, A.A.,
and Matthiopoulos, J. 2018. Used-habitat calibration plots: A new procedure
for validating species distribution, resource selection, and step-selection
models. \emph{Ecography} 41:737–752.
}
\seealso{
See Fieberg \emph{et al.} 2018 for details about UHC plots.

Default plotting method available: \code{\link{plot.uhc_sim}()}
}
