% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_steps.R, R/random_steps2.R
\name{random_steps}
\alias{random_steps}
\alias{random_steps.steps_xy}
\alias{random_steps2}
\alias{random_steps2.steps_xy}
\title{Generate Random Steps}
\usage{
random_steps(x, ...)

\method{random_steps}{steps_xy}(x, n_control = 10, sl_distr = "gamma",
  ta_distr = "vonmises", random.error = 0.001, ...)

random_steps2(x, ...)

\method{random_steps2}{steps_xy}(x, n_control = 10,
  sl_distr = fit_distr(sl_, "gamma"), ta_distr = fit_distr(ta_,
  "vonmises"), include_observed = TRUE, remove_first = TRUE, ...)
}
\arguments{
\item{x}{Steps.}

\item{...}{Further arguments, none implemented.}

\item{n_control}{\code{[integer(1)=10]{>1}} \cr The number of control steps paired with each observed step.}

\item{sl_distr}{\code{[character(1)='gamma']{'gamma'}} \cr The distribution to be fitted to the empirical distribution of step lengths.}

\item{ta_distr}{\code{[character(1)='vonmises']{'vonmises', 'unif'}} \cr The distribution to be fitted to the empirical distribution of turn angles.}

\item{random.error}{\code{[numeric(1)=0.001]{>0}} \cr Upper limit for a uniformly distributed random error
(between 0 and \code{random.error}) to be added to step lengths, to avoid step
lengths of length 0.}

\item{x}{Steps.}

\item{n_control}{\code{[integer(1)=10]{>1}} \cr The number of control steps paired
with each observed step.}

\item{sl_distr}{\code{[amt_distr]} \cr The step-length distribution.}

\item{ta_distr}{\code{[amt_distr]} \cr The turn-angle distribution.}

\item{random.error}{\code{[numeric(1)=0.001]{>0}} \cr Upper limit for a uniformly}

\item{...}{Further arguments, none implemented.}
}
\description{
Function to generate a given number of random steps for each observed step.

Function to generate a given number of random steps for each observed step.
}
