% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_steps.R
\name{random_steps}
\alias{random_steps}
\alias{random_steps.steps}
\title{Generate Random Steps}
\usage{
random_steps(x, ...)

\method{random_steps}{steps}(x, n_controll = 10, sl.distr = "gamma",
  ta.distr = "vonmises", random.error = 0.001, .progress = FALSE, ...)
}
\arguments{
\item{x}{Steps.}

\item{...}{Further arguments, none implemented.}

\item{n_controll}{The number of controll steps for each observed step.}

\item{sl.distr}{The distribution to be fitted to step lengths.}

\item{ta.distr}{The distribution to be fitted to the turn angles.}

\item{random.error}{Upper bount for a uniformly distributed random error
(beteen 0 and \code{random.error}) to be added to step lenghts, to avoid step
lengths of length 0.}

\item{.progress}{Should progress message be printed?}
}
\description{
Function to generate a given number of random steps for each observed stepl
}
