% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amt.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{filter}
\alias{group_by}
\alias{select}
\alias{summarise}
\alias{summarize}
\alias{ungroup}
\alias{distinct}
\alias{mutate}
\alias{pull}
\alias{hours}
\alias{minutes}
\alias{seconds}
\alias{days}
\alias{weeks}
\alias{map}
\alias{map_int}
\alias{NA_crs_}
\alias{clogit}
\alias{strata}
\alias{nest}
\alias{unnest}
\alias{tibble}
\alias{Surv}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:summarise]{summarize}}, \code{\link[dplyr:group_by]{ungroup}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{lubridate}{\code{\link[lubridate:period]{days}}, \code{\link[lubridate:period]{hours}}, \code{\link[lubridate:period]{minutes}}, \code{\link[lubridate:period]{seconds}}, \code{\link[lubridate:period]{weeks}}}

  \item{purrr}{\code{\link[purrr]{map}}, \code{\link[purrr:map]{map_int}}}

  \item{sf}{\code{\link[sf:st_crs]{NA_crs_}}}

  \item{survival}{\code{\link[survival]{clogit}}, \code{\link[survival]{strata}}, \code{\link[survival]{Surv}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}

  \item{tidyr}{\code{\link[tidyr]{nest}}, \code{\link[tidyr]{unnest}}}
}}

