% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_kernel.R, R/movement_kernel.R,
%   R/simulate_ud.R
\name{habitat_kernel}
\alias{habitat_kernel}
\alias{movement_kernel}
\alias{sim_ud}
\alias{simulate_ud}
\alias{issf_to_ud}
\title{Simulate UD from fitted SSF}
\usage{
habitat_kernel(fit, resources, exp = TRUE)

movement_kernel(fit, resources, quant = 0.99, adjust = FALSE)

simulate_ud(movement_kernel, habitat_kernel, start, n = 100000L)

issf_to_ud(fit, resources, start, quant = 0.99, adjust = TRUE, n = 1e+05)
}
\arguments{
\item{fit}{A fitted (i)SSF.}

\item{resources}{A `RasterLayer` or `RasterStack`, the resources available to the animals.}

\item{exp}{A logical scalar, indicating whether or not the resulting habitat kernel should be exponentiated. This is usually \code{TRUE}.}

\item{quant}{A numeric scalar, quantile of the step-length distribution that is
the maximum movement distance.}

\item{adjust}{A logical scalar, indicating whether or not step length
parameters should be adjusted or not. Note, parameter adjustment is only
possilbe, if movement related covariates were included in the model (e.g.,
step length and the log thereof).}

\item{movement_kernel}{A \code{RasterLayer} providing the movement kernel.}

\item{habitat_kernel}{A \code{RasterLayer} providing the kernel kernel.}

\item{start}{A numeric vector of length two, giving the starting point of the simulatoin.}

\item{n}{An integer, providing the number of simulation steps.}
}
\value{
The habitat kernel, as \code{RasterLayer}.
}
\description{
Function to obtain a habitat kernel from a fitted (i)SSF.
}
\details{
\strong{\code{movement_kernel()}:} calculates a movement kernel from a fitted
(i)SSF. The method is currently only implemented for the gamma
distribution.

The habitat kernel is calculated by multiplying resources with their corresponding coefficients from the fitted (i)SSF.

\strong{\code{simulate_ud()}:} simulates a utilization distribution (UD) from a fitted (iSSF).

\strong{\code{issf_to_ud()}:} this function is a convenience wrapper, that combines the functions \code{movement_kernel}, \code{habitat_kernel} and \code{simulate_ud}.
}
\references{
\insertRef{avgar2016}{amt}

\insertRef{signer2017}{amt}
}
