% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{sl_params}
\alias{sl_params}
\alias{sl_params.fit_clogit}
\alias{sl_params.random_steps}
\alias{sl_params.fitdist}
\title{Step lengths parameters}
\usage{
sl_params(x, ...)

\method{sl_params}{fit_clogit}(x, ...)

\method{sl_params}{random_steps}(x, ...)

\method{sl_params}{fitdist}(x, ...)
}
\arguments{
\item{x}{An object that contains either a fitted conditional logistic regression, random steps or just a fitted distribution.}

\item{...}{Further arguments, none implemented.}
}
\description{
Returns the parameter of the distribution (e.g., gamma) fitted to the distribution of step lengths.
}
