% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_methods.R
\name{step_length}
\alias{step_length}
\alias{step_lengths}
\alias{step_lengths.track_xy}
\alias{step_lengths_sq}
\alias{step_lengths_sq.track_xy}
\title{Step lengths}
\usage{
step_lengths(x, ...)

\method{step_lengths}{track_xy}(x, lonlat = FALSE, ...)

step_lengths_sq(x, ...)

\method{step_lengths_sq}{track_xy}(x, ...)
}
\arguments{
\item{x}{A track.}

\item{lonlat}{Logcial scalar, if \code{TRUE} geogrphic distances are calculated}
}
\description{
Funcitons to calculate step lenghts
}
\examples{
step_lengths(trk)
step_lengths(trk, lonlat = TRUE)

m <- move::move(x, y, now() + hours(1:10), proj = CRS("+init=epsg:4326"))
move::distance(m)

}
