% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{steps}
\alias{steps}
\alias{steps_by_burst}
\alias{steps_by_burst.track_xyt}
\alias{steps}
\alias{steps.track_xy}
\alias{steps.track_xyt}
\title{Track to Steps.}
\usage{
steps_by_burst(x, ...)

\method{steps_by_burst}{track_xyt}(x, ...)

steps(x, ...)

\method{steps}{track_xy}(x, ...)

\method{steps}{track_xyt}(x, ...)
}
\arguments{
\item{x}{A track.}

\item{...}{Further arguments, none implemented.}
}
\description{
Converts a track to steps.
}
