% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_align.R
\name{track_align}
\alias{track_align}
\alias{track_align.track_xyt}
\title{Selects relocations that fit a new time series}
\usage{
track_align(x, ...)

\method{track_align}{track_xyt}(x, nt, tol, ...)
}
\arguments{
\item{x}{A track.}

\item{...}{Further arguments, none implemented.}

\item{nt}{The new time trajectory.}

\item{tol}{The tolerance.}
}
\description{
Functions to only selects relocations that can be aligned with a new time series (within some tolerance).
}
