% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhc_plots.R
\name{ua_distr}
\alias{ua_distr}
\title{Summarize distribution of used and available}
\usage{
ua_distr(name, type, data, resp, avail = TRUE)
}
\arguments{
\item{name}{\verb{[character]} Name of the column to summarize.}

\item{type}{\verb{[character]} Either \code{"numeric"} or \code{"factor"} as returned by
\code{\link{prep_test_dat}()}.}

\item{data}{\verb{[data.frame]} The \code{data.frame} containing the columns and the
response variable.}

\item{resp}{\verb{[character]} Name of the response variable.}

\item{avail}{\verb{[logical]} Should distribution be calculated for the available
locations? Defaults to \code{TRUE}, but should be false when summarizing the
bootstrapped "used" samples.}
}
\description{
Summarizes distribution of numeric or factor variables
}
